### 2024-03-30 #6558 @editorialbot
[PRE REVIEW]: weatherOz: An API Client for Australian Weather and Climate Data Resources in R
**Submitting author:** @bozaah (<a href="http://orcid.org/0000-0001-7384-6849">Rodrigo Pires</a>)
**Repository:** https://github.com/ropensci/weatherOz
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
