### 2024-04-09 6cab03b
Upated DPIRD variable names and API url
