+++ R/get_dpird_summaries.R
@@ -103,6 +103,7 @@
 #'   * observationsCount,
 #'   * observationsPercentage,
 #'   * panEvaporation,
+#'   * panEvaporation12AM,
 #'   * rainfall,
 #'   * relativeHumidity,
 #'   * relativeHumidityAvg,
@@ -460,6 +461,7 @@ get_dpird_summaries <- function(station_code,
       "erosion_condition_minutes",
       "erosion_condition_start_time",
       "errors",
+      "evapotranspiration",
       "evapotranspiration_short_crop",
       "evapotranspiration_tall_crop",
       "frost_condition_minutes",
@@ -469,6 +471,7 @@ get_dpird_summaries <- function(station_code,
       "observations_count",
       "observations_percentage",
       "pan_evaporation",
+      "pan_evaporation_12am",
       "rainfall",
       "relative_humidity_avg",
       "relative_humidity_max",
@@ -524,6 +527,7 @@ get_dpird_summaries <- function(station_code,
       "erosion_condition_minutes",
       "erosion_condition_start_time",
       "errors",
+      "et",
       "et_short_crop",
       "et_tall_crop",
       "frost_condition_minutes",
@@ -533,6 +537,7 @@ get_dpird_summaries <- function(station_code,
       "observations_count",
       "observations_percentage",
       "pan_evaporation",
+      "pan_evaporation_12am",
       "rainfall",
       "rh_avg",
       "rh_tmax",
