+++ DESCRIPTION
@@ -122,7 +122,7 @@ Description: Provides automated downloading, parsing and formatting of weather
     data for Australia through API endpoints provided by the Department of
     Primary Industries and Regional Development (DPIRD) of Western Australia
     and by the Science and Technology Division of the Queensland Government's
-    Department of Environment and Science (DES).  As well as the Bureau of
+    Department of Environment and Science (DES). As well as the Bureau of
     Meteorology (BOM) of the Australian government precis and coastal
     forecasts, agriculture bulletin data, and downloading and importing radar
     and satellite imagery files.  DPIRD weather data are accessed through public
@@ -131,11 +131,11 @@ Description: Provides automated downloading, parsing and formatting of weather
     network.  Australia-wide weather data are based on data from the Australian
     Bureau of Meteorology (BOM) data and accessed through SILO (Scientific
     Information for Land Owners) (Jeffery et al., 2001,
-    <doi: 10.1016/S1364-8152(01)00008-1>).  DPIRD data are made available under
+    <doi: 10.1016/S1364-8152(01)00008-1>). DPIRD data are made available under
     a Creative Commons Attribution 3.0 Licence (CC BY 3.0 AU) license
-    <https://creativecommons.org/licenses/by/3.0/au/deed.en>.  SILO data are
+    <https://creativecommons.org/licenses/by/3.0/au/deed.en>. SILO data are
     released under a  Creative Commons Attribution 4.0 International licence
-    (CC BY 4.0) <https://creativecommons.org/licenses/by/4.0/>.  BOM data are
+    (CC BY 4.0) <https://creativecommons.org/licenses/by/4.0/>. BOM data are
     (c) Australian Government Bureau of Meteorology and released under a
     Creative Commons (CC) Attribution 3.0 licence or Public Access Licence (PAL)
     as appropriate, see <http://www.bom.gov.au/other/copyright.shtml> for
