### 2024-04-19 1fc11d4
Add CRAN badge to README
