+++ vignettes/use_case.Rmd
@@ -16,7 +16,7 @@ vignette: >
 
 ## Using {weatherOz} with the {WINS} for pulses project
 
-[Summer pulses (*e.g.*, mungbeans) are a high value crop for Queensland grain growers](http://www.pulseaus.com.au/about/australian-pulse-industry).
+[Summer pulses (*e.g.*, mungbeans) are a high value crop for Queensland grain growers](https://www.pulseaus.com.au/about/australian-pulse-industry).
 However, temperatures above 40 °C, particularly during flowering, can have a severe detrimental effect on crop yields (Kaushal *et al.* 2016).
 A key strategy that plants use to minimise the impact of heat stress is evaporative cooling.
 Ensuring crops are well watered prior to a heat stress event (so to facilitate high rates of evaporative cooling) is a potential strategy to ensure that irrigated crops avoid heat stress.
