+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -158,45 +158,45 @@ library(weatherOz)
   )
 )
 #> Key: <station_code>
-#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance
-#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>
-#>    1:           NO 2024-03-10 10:04:00            22.4              42.3             27.5              824
-#>    2:           NO 2024-03-10 10:05:00            22.5              41.3             27.6              826
-#>    3:           NO 2024-03-10 10:06:00            22.5              41.5             27.6              828
-#>    4:           NO 2024-03-10 10:07:00            22.6              41.2             27.7              831
-#>    5:           NO 2024-03-10 10:08:00            22.5              41.5             27.7              834
-#>   ---                                                                                                     
-#> 1423:           NO 2024-03-11 09:46:00            22.2              55.5             27.2              753
-#> 1424:           NO 2024-03-11 09:47:00            22.2              55.6             27.3              756
-#> 1425:           NO 2024-03-11 09:48:00            22.2              55.7             27.3              760
-#> 1426:           NO 2024-03-11 09:49:00            22.4              55.4             27.4              762
-#> 1427:           NO 2024-03-11 09:50:00            22.7              54.3             27.5              765
-#>       rainfall dew_point wet_bulb wind_height wind_avg_speed wind_avg_direction_compass_point
-#>          <int>     <num>    <num>       <int>          <num>                           <char>
-#>    1:        0       9.0     15.3           3           5.18                              ESE
-#>    2:        0       8.7     15.2           3           5.69                              ESE
-#>    3:        0       8.8     15.3           3           4.68                              ESE
-#>    4:        0       8.8     15.3           3           4.18                                E
-#>    5:        0       8.8     15.3           3           7.20                               NE
-#>   ---                                                                                        
-#> 1423:        0      12.9     17.1           3          10.73                               SE
-#> 1424:        0      12.9     17.1           3           8.21                               SE
-#> 1425:        0      12.9     17.1           3          10.73                              SSE
-#> 1426:        0      13.0     17.2           3           8.21                              SSE
-#> 1427:        0      13.0     17.3           3           5.69                              SSE
-#>       wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>                            <int>          <num>          <num>
-#>    1:                        105          2.664           8.21
-#>    2:                        117          3.672           9.22
-#>    3:                        107          1.656           7.20
-#>    4:                         91          1.152          10.22
-#>    5:                         54          4.680          10.73
-#>   ---                                                         
-#> 1423:                        131          5.184          16.24
-#> 1424:                        128          2.664          13.72
-#> 1425:                        150          6.696          16.74
-#> 1426:                        153          6.192          11.74
-#> 1427:                        154          4.176           7.20
+#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall
+#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <int>
+#>    1:           NO 2024-04-18 13:05:00            30.7              23.5             35.1              761        0
+#>    2:           NO 2024-04-18 13:06:00            31.0              23.1             35.1              763        0
+#>    3:           NO 2024-04-18 13:07:00            30.9              23.1             35.2              761        0
+#>    4:           NO 2024-04-18 13:08:00            30.9              23.2             35.2              761        0
[…]

[…]
