+++ DESCRIPTION
@@ -139,7 +139,7 @@ Description: Provides automated downloading, parsing and formatting of weather
     as appropriate, see <http://www.bom.gov.au/other/copyright.shtml> for
     further details.
 URL: https://github.com/ropensci/weatherOz/, https://docs.ropensci.org/weatherOz/
-BugReports: https://github.com/ropensci/weatherOz/issues
+BugReports: https://github.com/ropensci/weatherOz/issues/
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "rd",
