+++ README.Rmd
@@ -41,7 +41,7 @@ Follow the API Terms and Conditions for the [DPIRD](https://www.agric.wa.gov.au/
 
 Observation data from the DPIRD's weather station network is also available via a [web interface](https://weather.agric.wa.gov.au).
 The data available is a mirror of the DPIRD Weather 2.0 API endpoints.
-Rainfall estimates are also available at virtual stations (_i.e._, where no observational data is present) and is sourced from the Doppler radar service provided by the Bureau of Meteorology under license.
+Rainfall estimates are also available at virtual stations (_i.e._, where no observational data is present) and is sourced from the Doppler radar service provided by the Australian Government Bureau of Meteorology (BOM) under license.
 
 ## Installation instructions
 
@@ -50,6 +50,7 @@ You can install the development version of {weatherOz} like so:
 ```r
 install.packages("weatherOz", repos = "https://ropensci.r-universe.dev")
 ```
+
 ## A Note on API Keys
 
 The examples in this README assume that you have stored your API key in your .Renviron file.
