### 2024-04-19 a52e334
Update install instructions to include (aspirational) CRAN install instructions
