### 2024-04-21 #6558 @adamhsparks
@arfon, <https://github.com/ropensci/software-review/issues/598>
