### 2024-05-03 #6558 @adamhsparks
@arfon, I've fixed the broken references, they were in the bib, just typos, and removed the errant "[]" occurrences in the latest commits.
