### 2024-05-03 #6717 @bozaah
Thank you @arfon for moving our submission to review. Also, much appreciated @JanLauGe & @rogerssam for your time to 'continue' the review process here.

I am linking [issue #67](https://github.com/ropensci/weatherOz/issues/67) in the target repository that will contain replies and lengthier discussions from the review process.
