### 2024-05-03 b9c0500
Fixes https://github.com/ropensci/weatherOz/issues/69
