### 2024-05-03 2bdc947
normalise DESCRIPTION file, thanks Maelle!
