+++ DESCRIPTION
@@ -2,149 +2,65 @@ Package: weatherOz
 Title: An API Client for Australian Weather and Climate Data Resources
 Version: 1.0.0
 Authors@R: c(
-  person(
-    given = "Rodrigo",
-    family = "Pires",
-    role = c("aut", "cre"),
-    email = "rodrigo.pires@dpird.wa.gov.au",
-    comment = c(ORCID = "0000-0001-7384-6849")
-  ),
-  person(
-    given = "Anna",
-    family = "Hepworth",
-    role = "aut",
-    email = "anna.hepworth@gmail.com",
-    comment = c(ORCID = "0000-0003-0204-6347")
-  ),
-  person(
-    given = "Rebecca",
-    family = "O'Leary",
-    role = "aut",
-    email = "bec.oleary@curtin.edu.au"
-  ),
-  person(
-    given = "Jonathan",
-    family = "Carroll",
-    role = "aut",
-    email = "rpkg@jcarroll.com.au",
-    comment = c(ORCID = "0000-0002-1404-5264")
-  ),
-  person(
-    given = "James",
-    family = "Goldie",
-    role = "aut",
-    email = "me@jamesgoldie.dev",
-    comment = c(ORCID = "0000-0002-5024-6207")
-  ),
-  person(
-    given = "Dean",
-    family = "Marchiori",
-    role = "aut",
-    email = "deanmarchiori@gmail.com",
-    comment = c(ORCID = "0000-0002-3430-7225")
-  ),
-  person(
-    given = "Paul",
-    family = "Melloy",
-    role = "aut",
-    email = "paul@melloy.com.au",
-    comment = c(ORCID = "0000-0003-4253-7167")
-  ),
-  person(
-    given = "Mark",
-    family = "Padgham",
-    role = "aut",
-    email = "mark.padgham@email.com",
-    comment = c(ORCID = "0000-0003-2172-5265")
-  ),
-  person(
-    given = "Hugh",
-    family = "Parsonage",
-    role = "aut",
-    email = "hugh.parsonage@gmail.com",
-    comment = c(ORCID = "0000-0003-4055-0835")
-  ),
-  person(
-    given = "Keith",
-    family = "Pembleton",
-    role = "aut",
-    email = "keith.pembleton@usq.edu.au",
-    comment = c(ORCID = "0000-0002-1896-4516")
-  ),
-  person(
-    given = "Maëlle",
-    family = "Salmon",
-    role = "ctb",
-    email = "maelle.salmon@yahoo.se",
-    comment = c(ORCID = "0000-0002-2815-0399",
-                "Contributed to debugging a nasty little bug with CI where
-                timezones caused tests to fail due to vcr not recognising the
-                URL when run outside of Australia/Perth TZ! Suggested the use of
-                `withr::local_timzeone()`, see
-                <https://github.com/ropensci/weatherOz/commit/b052bf91973b8d7e147a39e8938405a64622634b>.")),
-  person(
-    given = "Jimmy",
-    family = "Ng",
-    role = "ctb",
-    email = "jimmy.ng@dpird.wa.gov.au"
-  ),
-  person(
-    given = "Steve",
-    family = "Collins",
-    role = "ctb",
-    email = "steve.collins@dpird.wa.gov.au",
-    comment = c("Designed weatherOz's hex logo.")
-  ),
-  person(
-    given = "Adam H.",
-    family = "Sparks",
-    email = "adamhsparks@gmail.com",
-    role = "aut",
-    comment = c(ORCID = "0000-0002-0061-8359")
-  ),
-  person(
-    given = "Laurens",
-    family = "Geffert",
-    role = "rev",
-    email = "laurensgeffert@gmail.com"
-  ),
[…]
