### 2024-05-03 b7d1b4d
redocument and rebuild codemeta.json after normalising DESCRIPTION file
