### 2024-05-04 5538d5a
Update README with JOSS badge
