### 2024-05-04 #6717 @arfon
@bozaah – please take a look at the compiled PDF linked above and make fixes where you can. There are a number of issues with formatting and some of the references not rendering properly (likely a mis-match between your BibTeX keys in the `paper.md` and `paper.bib` files).

You can check if things are fixed by asking @editorialbot for a new draft with `@editorialbot generate pdf` as a standalone comment in this thread.
