### 2024-05-09 #6717 @arfon
@JanLauGe – thanks for your enthusiasm here, but as I mentioned in the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/6558#issuecomment-2091533013), there's no need to do an additional review here (JOSS accepts the rOpenSci review).

I'm currently just waiting on @bozaah's response to[ this comment ](https://github.com/openjournals/joss-reviews/issues/6717#issuecomment-2094091767)to proceed.
