### 2024-05-14 0b3581f
Prefix amp_apsim_met() w/ '.' to not clobber fn from {apsimx}
