### 2024-05-15 8be64a1
Clarify documentation for .amp_apsim_met()
