### 2024-05-15 b5d618a
Simplify creation of DPIRD APSIM objects
