+++ R/get_dpird_apsim.R
@@ -1,5 +1,6 @@
 
 
+
 #' Get DPIRD Summary Weather Data in the APSIM Format From the Weather 2.0 API
 #'
 #' Automates the retrieval and conversion of summary data from the
@@ -18,6 +19,8 @@
 #' @param api_key A `character` string containing your \acronym{API} key from
 #'   \acronym{DPIRD}, <https://www.agric.wa.gov.au/web-apis>, for the
 #'   \acronym{DPIRD} Weather 2.0 \acronym{API}.
+#' @param filename A `character` string with your desired filename in the .met
+#'   object. Default \sQuote{noname.met} as with [apsimx::as_apsim_met()].
 #'
 #' @section Saving objects:
 #' To save \dQuote{met} objects, please use [apsimx::write_apsim_met()].
@@ -29,8 +32,8 @@
 #'
 #' wd <- get_dpird_apsim(
 #'   station_code = "BI",
-#'   start_date = "20220401",
-#'   end_date = "20221101",
+#'   start_date = "20220101",
+#'   end_date = "20221231",
 #'   api_key = "your_api_key"
 #' )
 #' }
@@ -38,8 +41,7 @@
 #'
 #' @author Adam H. Sparks, \email{adamhsparks@@gmail.com}
 #'
-#' @return An \CRANpkg{apsimx} object of class \sQuote{met} with attributes,
-#'   alternatively, a local APSIM .met file if `file` is specified.
+#' @return An \CRANpkg{apsimx} object of class \sQuote{met} with attributes.
 #'
 #' @family DPIRD
 #' @family data fetching
@@ -51,7 +53,8 @@
 get_dpird_apsim <- function(station_code,
                             start_date,
                             end_date = Sys.Date(),
-                            api_key) {
+                            api_key,
+                            filename = NULL) {
   apsim <- get_dpird_summaries(
     station_code = station_code,
     start_date = start_date,
@@ -69,22 +72,11 @@ get_dpird_apsim <- function(station_code,
     api_key = api_key
   )
 
-  station_name <- apsim$station_name[1]
-  longitude <- apsim$longitude[1]
-  latitude <- apsim$latitude[1]
   site <- apsim$station_name[1]
-  tav <- round(mean(colMeans(apsim[, c("air_tmax", "air_tmin")], na.rm = TRUE), na.rm = TRUE), 2)
-  tav <-
-    sprintf(
-      "tav = %g (oC) ! Average ambient temperature. Based on %s to %s.",
-      tav,
-      lubridate::ymd(start_date),
-      lubridate::ymd(end_date)
-    )
-
+  latitude = apsim$latitude[1]
+  longitude = apsim$longitude[1]
   apsim[, day := NULL]
   apsim[, day := lubridate::yday(apsim$date)]
-
   apsim <-
     apsim[, c(
       "year",
@@ -112,95 +104,31 @@ get_dpird_apsim <- function(station_code,
     new = c("radn", "maxt", "mint", "rain", "evap", "rh", "windspeed")
   )
 
-  data.table::setDF(apsim)
+  # if no `filename` is provided,
+  #. remove & use the default that `as_apsim_met()` provides, 'noname.met'
+  if (is.null(filename)) {
+    filename <- "noname.met"
+  }
 
-  units <-
-    c("()",
-      "()",
-      "(MJ/m2/day)",
-      "(oC)",
-      "(oC)",
-      "(mm)",
-      "(mm)",
-      "(%)",
-      "(m/s)")
-  comments <-
-    sprintf("!data from DPIRD Weather 2.0 API. retrieved: %s", Sys.time())
-
-  attr(apsim, "site") <- sprintf("%s.met", site)
[…]
