### 2024-05-15 191fdea
Use entire year for APSIM examples
