+++ R/get_data_drill_apsim.R
@@ -81,8 +81,8 @@
 #' wd <- get_data_drill_apsim(
 #'   latitude = -27.85,
 #'   longitude = 150.05,
-#'   start_date = "20220401",
-#'   end_date = "20221101",
+#'   start_date = "20220101",
+#'   end_date = "20221231",
 #'   api_key = "your_api_key"
 #' )
 #' }
