+++ R/get_patched_point_apsim.R
@@ -93,8 +93,8 @@
 #' # Source observation data for station Wongan Hills station, WA (008137)
 #' wd <- get_patched_point_apsim(
 #'   station_code = "008137",
-#'   start_date = "20220401",
-#'   end_date = "20221101",
+#'   start_date = "20220101",
+#'   end_date = "20221231",
 #'   api_key = "your_api_key"
 #' )
 #' }
