+++ man/get_data_drill_apsim.Rd
@@ -99,8 +99,8 @@ To save \dQuote{met} objects, please use \code{\link[apsimx:write_apsim_met]{aps
 wd <- get_data_drill_apsim(
   latitude = -27.85,
   longitude = 150.05,
-  start_date = "20220401",
-  end_date = "20221101",
+  start_date = "20220101",
+  end_date = "20221231",
   api_key = "your_api_key"
 )
 }
