### 2024-05-15 7e7d5a5
Use `as_apsim_met()` to import SILO data and add comments back
