+++ tests/testthat/test-get_dpird_apsim.R
@@ -1,14 +1,14 @@
 
 test_that("get_dpird_apsim() returns a 'met' object",
           {
-            vcr::use_cassette("dpird_apsim", {
+            vcr::use_cassette("dpird_get_apsim", {
               skip_if_offline()
-              x <- get_dpird_apsim(
+              expect_warning(x <- get_dpird_apsim(
                 station_code = "BI",
-                start_date = "20170228",
-                end_date = "20170228",
+                start_date = "20170101",
+                end_date = "20171231",
                 api_key = Sys.getenv("DPIRD_API_KEY")
-              )
+              ))
             })
             expect_s3_class(x, "met")
             expect_length(x, 9)
