+++ man/get_patched_point_apsim.Rd
@@ -9,7 +9,8 @@ get_patched_point_apsim(
   station_code,
   start_date,
   end_date = Sys.Date(),
-  api_key
+  api_key,
+  filename
 )
 }
 \arguments{
@@ -24,9 +25,12 @@ beginning of the range to query in the format \dQuote{yyyy-mm-dd}
 the range query in the format  \dQuote{yyyy-mm-dd} (ISO8601).  Data
 returned is inclusive of this date.  Defaults to the current system date.}
 
-\item{api_key}{A \code{character} string specifying a valid email address to use
+\item{api_key}{A \code{character} string providing a valid email address to use
 for the request.  The query will return an error if a valid email address
 is not provided.}
+
+\item{filename}{A \code{character} string providing an optional value for the
+\dQuote{filename} field in the .met file.}
 }
 \value{
 An \CRANpkg{apsimx} object of class \sQuote{met} with attributes.
@@ -105,7 +109,7 @@ not be printed as a part of the resulting \code{met} object in your \R session.
 \examples{
 \dontrun{
 # requires an API key as your email address
-# Source observation data for station Wongan Hills station, WA (008137)
+# Source observation data for Wongan Hills station, WA (008137)
 wd <- get_patched_point_apsim(
   station_code = "008137",
   start_date = "20220101",
