### 2024-05-16 9b5dc0b
Moves fn into reexports.R
