+++ man/get_dpird_apsim.Rd
@@ -5,13 +5,7 @@
 \alias{get_dpird_apsim}
 \title{Get DPIRD Summary Weather Data in the APSIM Format From the Weather 2.0 API}
 \usage{
-get_dpird_apsim(
-  station_code,
-  start_date,
-  end_date = Sys.Date(),
-  api_key,
-  filename = NULL
-)
+get_dpird_apsim(station_code, start_date, end_date = Sys.Date(), api_key)
 }
 \arguments{
 \item{station_code}{A \code{character} string of the \acronym{DPIRD} station code
@@ -29,9 +23,6 @@ returned is inclusive of this date.  Defaults to the current system date.}
 \item{api_key}{A \code{character} string containing your \acronym{API} key from
 \acronym{DPIRD}, \url{https://www.agric.wa.gov.au/web-apis}, for the
 \acronym{DPIRD} Weather 2.0 \acronym{API}.}
-
-\item{filename}{A \code{character} string with your desired filename in the .met
-object. Default \sQuote{noname.met} as with \code{\link[apsimx:as_apsim_met]{apsimx::as_apsim_met()}}.}
 }
 \value{
 An \CRANpkg{apsimx} object of class \sQuote{met} with attributes.
@@ -43,7 +34,9 @@ formatted weather data object.
 }
 \section{Saving objects}{
 
-To save \dQuote{met} objects, please use \code{\link[apsimx:write_apsim_met]{apsimx::write_apsim_met()}}.
+To save \dQuote{met} objects the \code{\link[apsimx:write_apsim_met]{apsimx::write_apsim_met()}} is reexported.
+Note that when saving, comments from SILO will be included, but these will
+not be printed as a part of the resulting \code{met} object in your \R session.
 }
 
 \examples{
@@ -89,7 +82,8 @@ Other data fetching:
 
 Other APSIM: 
 \code{\link{get_data_drill_apsim}()},
-\code{\link{get_patched_point_apsim}()}
+\code{\link{get_patched_point_apsim}()},
+\code{\link{reexports}}
 }
 \author{
 Adam H. Sparks, \email{adamhsparks@gmail.com}
