+++ man/get_data_drill_apsim.Rd
@@ -10,8 +10,7 @@ get_data_drill_apsim(
   latitude,
   start_date,
   end_date = Sys.Date(),
-  api_key,
-  filename
+  api_key
 )
 }
 \arguments{
@@ -32,9 +31,6 @@ returned is inclusive of this date.  Defaults to the current system date.}
 \item{api_key}{A \code{character }string specifying a valid email address to use
 for the request.  The query will return an error if a valid email address
 is not provided.}
-
-\item{filename}{A \code{character} string providing an optional value for the
-\dQuote{filename} field in the .met file.}
 }
 \value{
 An \CRANpkg{apsimx} object of class \sQuote{met} with attributes.
@@ -47,6 +43,10 @@ splining or kriging the observational data.  The returned values contain
 \dQuote{source} columns, which denote how the observations were derived.
 The grid spans 112° to 154°, -10° to -44° with resolution 0.05° latitude by
 0.05° longitude (approximately 5 km × 5 km).
+}
+\details{
+Note that when saving, comments from SILO will be included, but these will
+not be printed as a part of the resulting \code{met} object in your \R session.
 }
 \section{Included Values}{
 
@@ -92,7 +92,7 @@ interpolation method for \acronym{CLIMARC} data;}
 
 \section{Saving objects}{
 
-To save \dQuote{met} objects, please use \code{\link[apsimx:write_apsim_met]{apsimx::write_apsim_met()}}.
+To save \dQuote{met} objects the \code{\link[apsimx:write_apsim_met]{apsimx::write_apsim_met()}} is reexported.
 Note that when saving, comments from SILO will be included, but these will
 not be printed as a part of the resulting \code{met} object in your \R session.
 }
@@ -147,7 +147,8 @@ Other data fetching:
 
 Other APSIM: 
 \code{\link{get_dpird_apsim}()},
-\code{\link{get_patched_point_apsim}()}
+\code{\link{get_patched_point_apsim}()},
+\code{\link{reexports}}
 }
 \author{
 Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}, and Adam
