+++ NAMESPACE
@@ -24,6 +24,8 @@ export(parse_ag_bulletin)
 export(parse_coastal_forecast)
 export(parse_precis_forecast)
 export(plot)
+export(write_apsim_met)
+importFrom(apsimx,write_apsim_met)
 importFrom(data.table,":=")
 importFrom(data.table,.BY)
 importFrom(data.table,.EACHI)
