### 2024-05-19 7b298c8
add south_west_agricultural_region data and move docs to data.R
