+++ R/find_stations_in.R
@@ -1,0 +1,188 @@
+
+#' Find Stations Within a Geographic Area of Interest
+#'
+#' Given an \CRANpkg{sf} polygon, a bounding box as a vector with the minimum
+#'   and maximum longitude and latitude values or a place name, find
+#'   \acronym{DPIRD} or \acronym{BOM} stations in the \acronym{SILO} network
+#'   that fall within that defined area or the station nearest the centroid of
+#'   the area of interest.
+#'
+#' @param polygon An object of class \CRANpkg{sf} defining the area of interest.
+#' @param bbox `Vector` A four-digit vector defining a bounding box of the area
+#'  of interest in this order, \sQuote{xmin}, \sQuote{ymin}, \sQuote{xmax},
+#'   \sQuote{ymax}.
+#' @param place_name `Character` A string value containing the place name of the
+#'   area of interest to query from Open Street Map using [osmdata::getbb].
+#'   Results are the same as providing your own bounding box, a rectangular
+#'   bounding box, except that you don't need to know what the coordinates are
+#'   if you trust Open Street Map.
+#' @param centroid `Boolean` A value of `TRUE` or `FALSE` indicating whether
+#'   you want the centroid only to be used to find the nearest station to the
+#'   centre of the area of interest.  If \dQuote{n} polygons are supplied,
+#'   \dQuote{n} stations are returned.  Defaults to `FALSE` with all stations
+#'   within the area of interest returned.
+#' @param api_key A `string` value that is the user's \acronym{API} key from
+#'   \acronym{DPIRD} (see <https://www.agric.wa.gov.au/web-apis>).  Only used
+#'   when \var{which_api} is `DPIRD` or `all`.
+#' @param which_api A `string` value that indicates which \acronym{API} to use.
+#'   Defaults to `silo` only.  Valid values are `all`, for both \acronym{SILO}
+#'   (\acronym{BOM}) and \acronym{DPIRD} weather station networks; `silo` for
+#'   only stations in the \acronym{SILO} network; or `dpird` for stations in the
+#'   \acronym{DPIRD} network.
+#' @param include_closed A `Boolean` value that indicates whether closed
+#'   stations in the \acronym{DPIRD} network should be included in the results.
+#'   Defaults to `FALSE` with closed stations not included.
+#' @param crs A `string` value that provides the coordinate reference system,
+#'   AKA, "projection" to be used for the point extraction.  Defaults to
+#'   GDA 2020, EPSG:7844.  **NOTE** This will override any `crs` value that your
+#'.  `polygon` provides unless you specify it again here, *e.g.*,
+#'    `crs = sf::st_crs(polygon_object_name)`.
+#'
+#' @return a \CRANpkg{data.table} object of weather station(s) within the
+#'   defined area of interest in an unprojected format, EPSG:4326, WGS 84 --
+#'   WGS84 - World Geodetic System 1984, used in \acronym{GPS} format.
+#'
+#' @examples
+#' # using a named place, Toowoomba, Qld, AU using only the SILO API for BOM
+#' # stations
+#'
+#' find_stations_in(place_name = "Toowoomba Qld",
+#'                  which_api = "SILO",
+#'                  include_closed = TRUE)
+#'
+#' # using a bounding box for Melbourne, Vic using only the SILO API for BOM
+#' # stations
+#'
+#' find_stations_in(
+#'   bbox = c(144.470215, -38.160476, 145.612793, -37.622934),
+#'   api_key = "your_api_key",
+#'   which_api = "SILO",
+#'   include_closed = TRUE
+#' )
+#'
+#' # Use the same bounding box but only find a single station nearest
+#'# the centroid using only the SILO API for BOM stations
+#'
[…]
