+++ R/dpird_summary_values.R
@@ -1,17 +1,0 @@
-
-#' A List of DPIRD Summary Weather Data Values
-#'
-#' A [vector] object containing 75 items representing valid values to supply
-#'   to `get_dpird_summary()`'s \var{values} argument taken from the
-#'   documentation for the \acronym{DPIRD} Weather 2.0 \acronym{API}.
-#'
-#' @format A [vector] object of 75 items.
-#'
-#' @keywords datasets
-#'
-#' @family DPIRD
-#' @family data
-#'
-#' @source <https://www.agric.wa.gov.au/weather-api-20>
-#'
-"dpird_summary_values"
