+++ DESCRIPTION
@@ -75,6 +75,7 @@ Imports:
     lubridate,
     magick,
     methods,
+    osmdata,
     readr (>= 1.1.1),
     stars,
     stats,
