+++ man/dpird_summary_values.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dpird_summary_values.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{dpird_summary_values}
 \alias{dpird_summary_values}
@@ -23,6 +23,7 @@ Other DPIRD:
 \code{\link{dpird_extreme_weather_values}},
 \code{\link{dpird_minute_values}},
 \code{\link{find_nearby_stations}()},
+\code{\link{find_stations_in}()},
 \code{\link{get_dpird_apsim}()},
 \code{\link{get_dpird_availability}()},
 \code{\link{get_dpird_extremes}()},
