+++ man/get_dpird_apsim.Rd
@@ -60,6 +60,7 @@ Other DPIRD:
 \code{\link{dpird_minute_values}},
 \code{\link{dpird_summary_values}},
 \code{\link{find_nearby_stations}()},
+\code{\link{find_stations_in}()},
 \code{\link{get_dpird_availability}()},
 \code{\link{get_dpird_extremes}()},
 \code{\link{get_dpird_minute}()},
