+++ man/find_forecast_towns.Rd
@@ -46,6 +46,7 @@ Other BOM:
 
 Other metadata: 
 \code{\link{find_nearby_stations}()},
+\code{\link{find_stations_in}()},
 \code{\link{get_available_imagery}()},
 \code{\link{get_available_radar}()},
 \code{\link{get_dpird_availability}()},
