+++ man/find_stations_in.Rd
@@ -1,0 +1,151 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/find_stations_in.R
+\name{find_stations_in}
+\alias{find_stations_in}
+\title{Find Stations Within a Geographic Area of Interest}
+\usage{
+find_stations_in(
+  polygon = NULL,
+  bbox = NULL,
+  place_name = NULL,
+  centroid = FALSE,
+  api_key = "your_api_key",
+  which_api = "all",
+  include_closed = FALSE,
+  crs = "EPSG:7844"
+)
+}
+\arguments{
+\item{polygon}{An object of class \CRANpkg{sf} defining the area of interest.}
+
+\item{bbox}{\code{Vector} A four-digit vector defining a bounding box of the area
+of interest in this order, \sQuote{xmin}, \sQuote{ymin}, \sQuote{xmax},
+\sQuote{ymax}.}
+
+\item{place_name}{\code{Character} A string value containing the place name of the
+area of interest to query from Open Street Map using \link[osmdata:getbb]{osmdata::getbb}.
+Results are the same as providing your own bounding box, a rectangular
+bounding box, except that you don't need to know what the coordinates are
+if you trust Open Street Map.}
+
+\item{centroid}{\code{Boolean} A value of \code{TRUE} or \code{FALSE} indicating whether
+you want the centroid only to be used to find the nearest station to the
+centre of the area of interest.  If \dQuote{n} polygons are supplied,
+\dQuote{n} stations are returned.  Defaults to \code{FALSE} with all stations
+within the area of interest returned.}
+
+\item{api_key}{A \code{string} value that is the user's \acronym{API} key from
+\acronym{DPIRD} (see \url{https://www.agric.wa.gov.au/web-apis}).  Only used
+when \var{which_api} is \code{DPIRD} or \code{all}.}
+
+\item{which_api}{A \code{string} value that indicates which \acronym{API} to use.
+Defaults to \code{silo} only.  Valid values are \code{all}, for both \acronym{SILO}
+(\acronym{BOM}) and \acronym{DPIRD} weather station networks; \code{silo} for
+only stations in the \acronym{SILO} network; or \code{dpird} for stations in the
+\acronym{DPIRD} network.}
+
+\item{include_closed}{A \code{Boolean} value that indicates whether closed
+stations in the \acronym{DPIRD} network should be included in the results.
+Defaults to \code{FALSE} with closed stations not included.}
+
+\item{crs}{A \code{string} value that provides the coordinate reference system,
+AKA, "projection" to be used for the point extraction.  Defaults to
+GDA 2020, EPSG:7844.  \strong{NOTE} This will override any \code{crs} value that your
+.  \code{polygon} provides unless you specify it again here, \emph{e.g.},
+\code{crs = sf::st_crs(polygon_object_name)}.}
+}
+\value{
+a \CRANpkg{data.table} object of weather station(s) within the
+defined area of interest in an unprojected format, EPSG:4326, WGS 84 --
+WGS84 - World Geodetic System 1984, used in \acronym{GPS} format.
+}
+\description{
+Given an \CRANpkg{sf} polygon, a bounding box as a vector with the minimum
+and maximum longitude and latitude values or a place name, find
+\acronym{DPIRD} or \acronym{BOM} stations in the \acronym{SILO} network
+that fall within that defined area or the station nearest the centroid of
+the area of interest.
+}
+\examples{
+# using a named place, Toowoomba, Qld, AU using only the SILO API for BOM
+# stations
+
[…]
