+++ man/south_west_agriculture_region.Rd
@@ -1,0 +1,45 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{south_west_agriculture_region}
+\alias{south_west_agriculture_region}
+\alias{south_west_agricultural_region}
+\title{Western Australia Southwest Agriculture Region Geospatial Polygon}
+\format{
+An \code{\link[sf:sf]{sf::sf()}} polygon object
+}
+\source{
+\href{https://catalogue.data.wa.gov.au/dataset/south-west-agricultural-region-dpird-008}{Western Australia Department of Primary Industries and Regional
+Development}
+under a \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 Licence}
+}
+\usage{
+south_west_agricultural_region
+}
+\description{
+An \CRANpkg{sf} object of the the WA South West Agricultural Region.
+}
+\details{
+The zone managed for intensive agricultural activities in South-Western
+Australia.  Also known as the South West Agricultural Area or Clearing Line.
+This zone defines the easternmost extent of land cleared for agricultural
+purposes.
+}
+\section{Base data sets}{
+
+Western Australian Land Information Authority -
+Captured from photographic interpretation of best available orthophotography
+at date of capture, dates range between 2007 and 2010.
+}
+
+\section{Scale of capture}{
+
+1:20,000
+}
+
+\section{Coordinate Reference System}{
+
+EPSG:28350 - GDA94 / MGA zone 50 \url{https://epsg.io/28350}
+}
+
+\keyword{datasets}
