### 2024-05-19 1a9bef9
Simplify inputs to just be "x" for any of three object types
