+++ R/find_stations_in.R
@@ -7,15 +7,16 @@
 #'   that fall within that defined area or the station nearest the centroid of
 #'   the area of interest.
 #'
-#' @param polygon An object of class \CRANpkg{sf} defining the area of interest.
-#' @param bbox `Vector` A four-digit vector defining a bounding box of the area
-#'  of interest in this order, \sQuote{xmin}, \sQuote{ymin}, \sQuote{xmax},
-#'   \sQuote{ymax}.
-#' @param place_name `Character` A string value containing the place name of the
-#'   area of interest to query from Open Street Map using [osmdata::getbb].
-#'   Results are the same as providing your own bounding box, a rectangular
-#'   bounding box, except that you don't need to know what the coordinates are
-#'   if you trust Open Street Map.
+#' @param x One of three types of object:
+#'   * A `Vector` A four-digit vector defining a bounding box of the area
+#'   of interest in this order, \sQuote{xmin}, \sQuote{xmax}, \sQuote{ymin},
+#'   \sQuote{ymax},
+#'   * An object of class \CRANpkg{sf} defining the area of interest, or
+#'   * A `Character` string value containing the place name of the
+#'   area of interest to query from Open Street Map using [osmdata::getbb],
+#'   *e.g.*, \dQuote{Sydney, NSW}.  These results are the same as providing your
+#'   own bounding box, a rectangular bounding box, except that you don't need to
+#'   know what the coordinates are if you trust Open Street Map.
 #' @param centroid `Boolean` A value of `TRUE` or `FALSE` indicating whether
 #'   you want the centroid only to be used to find the nearest station to the
 #'   centre of the area of interest.  If \dQuote{n} polygons are supplied,
@@ -97,60 +98,34 @@
 #'
 #' @export
 
-find_stations_in <- function(polygon = NULL,
-                             bbox = NULL,
-                             place_name = NULL,
+find_stations_in <- function(x,
                              centroid = FALSE,
                              api_key = "your_api_key",
                              which_api = "all",
                              include_closed = FALSE,
                              crs = "EPSG:7844"
                              ) {
-  pars <- list(bbox = bbox,
-                 polygon = polygon,
-                 place_name = place_name)
-
-  if (all(vapply(pars, is.null, TRUE))) {
-    stop("You must provide one of `bbox`, `polygon` or `place_name`.")
-  }
-
-  if (length(pars[vapply(pars, is.null, TRUE)]) < 2) {
-    warning(
-      "You have provided multiple values for the area of interest. ",
-      "Only one will be used in this order, `polygon`, `bbox`, `place_name`."
-    )
-    if (!is.null(parspolygon)) {
-      area <- parspolygon
-    } else if (!is.null(parsbbox)) {
-      area <- parsbbox
-    } else {
-      area <- parsplace_name
-    }
-  } else {
-  # if we have just one parameter, we select the only non-null value
-  area <- Filter(Negate(is.null), pars)[[1]]
-  }
 
   # convert bbox or named places to {sf} polygons
-  if (is.numeric(area)) {
+  if (is.numeric(x)) {
     # area is bbox -----
-    area <- sf::st_as_sf(
-      data.table::data.table("x" = area[c(1, 3)], "y" = area[c(2, 4)]),
+    x <- sf::st_as_sf(
+      data.table::data.table("x" = x[c(1, 3)], "y" = x[c(2, 4)]),
       coords = c("x", "y"),
       crs = "EPSG:4356"
     )
-    area <- sf::st_as_sfc(sf::st_bbox(area), crs = crs)
-  } else if (is.character(area)) {
[…]
