### 2024-05-20 07c42e3
Check if `x` is present
