### 2024-05-20 128aa19
update tests to use `x` parameter
