### 2024-05-20 67a6c73
Adds tests and fixtures for new functionality
