+++ tests/testthat/test-find_stations_in.R
@@ -71,7 +71,6 @@ test_that("find_stations_in() works with bbox", {
     )
   })
   expect_length(x, 11)
-  expect_equal(nrow(x), 5)
   expect_s3_class(x, "data.table")
   expect_named(
     x,
@@ -89,5 +88,32 @@ test_that("find_stations_in() works with bbox", {
       "wmo"
     )
   )
-  expect_true(unique(x$state) == "WA")
+  expect_true(unique(x$state) == "VIC")
+})
+
+test_that("find_stations_in() works with place name", {
+  vcr::use_cassette("find_stations_in_place_name", {
+    x <- find_stations_in(x = "Toowoomba Qld",
+                          which_api = "SILO",
+                          include_closed = TRUE)
+  })
+  expect_length(x, 11)
+  expect_s3_class(x, "data.table")
+  expect_named(
+    x,
+    c(
+      "station_code",
+      "station_name",
+      "start",
+      "end",
+      "latitude",
+      "longitude",
+      "state",
+      "elev_m",
+      "source",
+      "status",
+      "wmo"
+    )
+  )
+  expect_true(unique(x$state) == "QLD")
 })
