+++ DESCRIPTION
@@ -67,7 +67,7 @@ Imports:
     crayon,
     crul,
     curl,
-    data.table,
+    data.table (>= 1.1.5),
     foreign,
     grDevices,
     jsonlite,
@@ -77,6 +77,7 @@ Imports:
     methods,
     osmdata,
     readr (>= 1.1.1),
+    sf,
     stars,
     stats,
     stringi,
