+++ R/weatherOz-package.R
@@ -12,6 +12,7 @@
 #' @importFrom data.table .N
 #' @importFrom data.table .NGRP
 #' @importFrom data.table .SD
+#' @importFrom data.table %notin%
 #' @importFrom data.table data.table
 ## usethis namespace: end
 NULL
