+++ NAMESPACE
@@ -27,6 +27,7 @@ export(parse_precis_forecast)
 export(plot)
 export(write_apsim_met)
 importFrom(apsimx,write_apsim_met)
+importFrom(data.table,"%notin%")
 importFrom(data.table,":=")
 importFrom(data.table,.BY)
 importFrom(data.table,.EACHI)
