### 2024-05-20 0b4aef8
Tidy docs and use examplesIf()
