+++ man/find_stations_in.Rd
@@ -17,7 +17,7 @@ find_stations_in(
 \item{x}{One of three types of object:
 \itemize{
 \item A \code{Vector} A four-digit vector defining a bounding box of the area
-of interest in this order, \sQuote{xmin}, \sQuote{xmax}, \sQuote{ymin},
+of interest in this order, \sQuote{xmin}, \sQuote{ymin}, \sQuote{xmax},
 \sQuote{ymax},
 \item An object of class \CRANpkg{sf} defining the area of interest, or
 \item A \code{Character} string value containing the place name of the
