### 2024-05-20 9efafbe
Add extra checks for bbox inputs
