### 2024-05-20 97fe1e9
Put latlon checks in proper place
