### 2024-05-20 de6fe39
Remove redundant code
