+++ vignettes/weatherOz_for_SILO.Rmd
@@ -94,7 +94,7 @@ The location as longitude and latitude, `longitude` and `latitude`, values must 
 Using coordinates for Southwood, Qld, get Data Drill data for all values available starting on 2022-01-01.
 
 
-```r
+``` r
 library(weatherOz)
 
 (
@@ -105,84 +105,123 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude  year month   day       date air_tmax air_tmax_source air_tmin air_tmin_source  elev_m
-#>          <num>    <num> <num> <num> <int>     <Date>    <num>           <int>    <num>           <int>  <char>
-#>   1:    150.05   -27.85  2022     1     1 2022-01-01     29.7              25     18.8              25 254.5 m
-#>   2:    150.05   -27.85  2022     1     2 2022-01-02     33.6              25     16.6              25 254.5 m
-#>   3:    150.05   -27.85  2022     1     3 2022-01-03     35.7              25     18.1              25 254.5 m
-#>   4:    150.05   -27.85  2022     1     4 2022-01-04     36.3              25     18.1              25 254.5 m
-#>   5:    150.05   -27.85  2022     1     5 2022-01-05     36.4              25     20.3              25 254.5 m
-#>  ---                                                                                                          
-#> 835:    150.05   -27.85  2024     4    14 2024-04-14     29.2              25     14.4              25 254.5 m
-#> 836:    150.05   -27.85  2024     4    15 2024-04-15     29.4              25     15.2              25 254.5 m
-#> 837:    150.05   -27.85  2024     4    16 2024-04-16     29.6              25     14.4              25 254.5 m
-#> 838:    150.05   -27.85  2024     4    17 2024-04-17     29.8              25     14.8              25 254.5 m
-#> 839:    150.05   -27.85  2024     4    18 2024-04-18     27.7              75     16.1              25 254.5 m
-#>      et_morton_actual et_morton_actual_source et_morton_potential et_morton_potential_source et_morton_wet
-#>                 <num>                   <int>               <num>                      <int>         <num>
-#>   1:              4.3                      26                 7.2                         26           5.7
-#>   2:              6.2                      26                 9.6                         26           7.9
-#>   3:              4.3                      26                11.5                         26           7.9
-#>   4:              5.1                      26                11.1                         26           8.1
-#>   5:              5.0                      26                10.9                         26           8.0
-#>  ---                                                                                                      
-#> 835:              2.8                      26                 5.6                         26           4.2
-#> 836:              2.7                      26                 5.8                         26           4.3
-#> 837:              2.7                      26                 5.4                         26           4.0
-#> 838:              3.2                      26                 5.4                         26           4.3
-#> 839:              1.7                      26                 4.0                         26           2.8
-#>      et_morton_wet_source et_short_crop et_short_crop_source et_tall_crop et_tall_crop_source evap_comb
-#>                     <int>         <num>                <int>        <num>               <int>     <num>
-#>   1:                   26           5.2                   26          6.3                  26       5.0
-#>   2:                   26           7.0                   26          8.6                  26       9.2
-#>   3:                   26           7.7                   26          9.9                  26       8.1
-#>   4:                   26           7.7                   26          9.7                  26       8.6
-#>   5:                   26           7.5                   26          9.4                  26       6.3
-#>  ---                                                                                                   
-#> 835:                   26           3.9                   26          5.0                  26       4.2
-#> 836:                   26           4.0                   26          5.1                  26       3.9
-#> 837:                   26           3.8                   26          4.9                  26       4.1
[…]
