+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -97,7 +97,7 @@ You must provide a `station_code` and `API_key`, the other arguments, `values` a
 In the first example, we illustrate how to fetch all extreme values available for Northam.
 
 
-```r
+``` r
 library(weatherOz)
 
 (extremes <- get_dpird_extremes(
@@ -112,7 +112,7 @@ Fetch only soil erosion extreme conditions for Northam, WA.
 The documentation for `get_dpird_extremes()` contains a full listing of the values that are available to query from this API endpoint.
 
 
-```r
+``` r
 library(weatherOz)
 
 (
@@ -148,7 +148,7 @@ You must provide a `station_code` and `API_key`, the other arguments, `start_dat
 ### Example 3: Get All Minute Data for the Past 24 Hours
 
 
-```r
+``` r
 library(weatherOz)
 
 (
@@ -158,45 +158,58 @@ library(weatherOz)
   )
 )
 #> Key: <station_code>
-#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall
-#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <int>
-#>    1:           NO 2024-04-18 13:05:00            30.7              23.5             35.1              761        0
-#>    2:           NO 2024-04-18 13:06:00            31.0              23.1             35.1              763        0
-#>    3:           NO 2024-04-18 13:07:00            30.9              23.1             35.2              761        0
-#>    4:           NO 2024-04-18 13:08:00            30.9              23.2             35.2              761        0
-#>    5:           NO 2024-04-18 13:09:00            31.1              22.9             35.2              761        0
-#>   ---                                                                                                              
-#> 1432:           NO 2024-04-19 12:56:00            28.7              22.9             33.3              776        0
-#> 1433:           NO 2024-04-19 12:57:00            28.6              23.0             33.4              774        0
-#> 1434:           NO 2024-04-19 12:58:00            28.5              23.0             33.4              773        0
-#> 1435:           NO 2024-04-19 12:59:00            28.3              23.4             33.5              772        0
-#> 1436:           NO 2024-04-19 13:00:00            28.2              23.3             33.5              773        0
-#>       dew_point wet_bulb wind_height wind_avg_speed wind_avg_direction_compass_point wind_avg_direction_degrees
-#>           <num>    <num>       <int>          <num>                           <char>                      <int>
-#>    1:       7.5     17.5           3           4.18                               NE                         54
-#>    2:       7.5     17.6           3           4.68                               NE                         34
-#>    3:       7.4     17.5           3           3.17                              NNE                         13
-#>    4:       7.5     17.5           3           5.69                              NNE                         26
-#>    5:       7.5     17.6           3           6.19                                E                         79
-#>   ---                                                                                                          
-#> 1432:       5.5     16.0           3           3.17                              ENE                         77
-#> 1433:       5.5     16.0           3           2.66                               NE                         52
[…]
