### 2024-05-21 a069298
Fix bug importing BOM metadata that prevented merging with SILO metadata
