### 2024-05-21 fce61c6
Check that user hasn't provided e-mail address for DPIRD API key
