+++ R/get_stations_metadata.R
@@ -130,9 +130,11 @@ get_stations_metadata <-
     } else if (which_api == "all") {
       if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
         stop(call. = FALSE,
-             "You must provide an API key for this query.")
+             "A valid DPIRD API key must be provided, please visit\n",
+             "<https://www.agric.wa.gov.au/web-apis> to request one.\n",)
       }
       .check_not_example_api_key(api_key)
+      .is_valid_dpird_api_key(api_key)
       silo <- .fetch_silo_metadata()
       dpird <- .fetch_dpird_metadata(.api_key = api_key, .rich = rich)
       out <- data.table::rbindlist(list(silo, dpird), fill = TRUE)
