### 2024-05-21 59ad494
Tidy .bib file, escape characters, ensure caps and fix duplicated keys
