+++ inst/paper/paper.bib
@@ -1,4 +1,4 @@
-@Webpage{BoM2024,
+@Webpage{BOM2024,
   author      = {{Australian Bureau of Meteorology}},
   lastchecked = {23/02/2024},
   month       = feb,
@@ -7,10 +7,10 @@
   year        = {2024},
 }
 
-@Article{Sparks2017,
+@Article{Sparks2017a,
   author    = {Sparks, Adam H. and Padgham, Mark and Parsonage, Hugh and Pembleton, Keith},
   journal   = {The Journal of Open Source Software},
-  title     = {{bomrang}: Fetch Australian Government Bureau of Meteorology Data in {R}},
+  title     = {{bomrang}: {Fetch Australian} Government {Bureau of Meteorology} Data in {R}},
   year      = {2017},
   issn      = {2475-9066},
   month     = sep,
@@ -22,7 +22,7 @@
 }
 
 @Manual{Miguez2024,
-  title  = {{apsimx}: Inspect, Read, Edit and Run {'APSIM' "Next Generation" }and {'APSIM' Classic}},
+  title  = {{apsimx}: Inspect, Read, Edit and Run {'APSIM' "Next Generation"} and {'APSIM' Classic}},
   author = {Miguez, Fernando},
   note   = {R package version 2.6.2},
   year   = {2024},
@@ -31,7 +31,7 @@
 
 @Article{Jeffrey2001,
   author    = {Jeffrey, Stephen J. and Carter, John O. and Moodie, Keith B. and Beswick, Alan R.},
-  journal   = {Environmental Modelling &amp; Software},
+  journal   = {Environmental Modelling \& Software},
   title     = {Using spatial interpolation to construct a comprehensive archive of {Australian} climate data},
   year      = {2001},
   issn      = {1364-8152},
@@ -105,7 +105,7 @@
   journal    = {Phytopathology},
   title      = {Risk Assessment Models for Wheat Fusarium Head Blight Epidemics Based on Within-Season Weather Data},
   year       = {2003},
-  month      = {2017/08/28},
+  month      = aug,
   number     = {4},
   pages      = {428--435},
   volume     = {93},
@@ -130,7 +130,7 @@
 }
 
 @Book{Pebesma2023,
-  author    = {Edzer Pebesma and Roger Bivand},
+  author    = {Pebesma, Edzer and Bivand, Roger},
   publisher = {Chapman and Hall/CRC},
   title     = {{Spatial Data Science: With applications in R}},
   year      = {2023},
@@ -142,7 +142,7 @@
 
 @Manual{Hijmans2024,
   title  = {{terra}: Spatial Data Analysis},
-  author = {Robert J. Hijmans},
+  author = {Hijmans, Robert J.},
   note   = {R package version 1.7-71},
   year   = {2024},
   url    = {https://CRAN.R-project.org/package=terra},
@@ -150,7 +150,7 @@
 
 @Manual{Barrett2024,
   title  = {{data.table}: Extension of `data.frame`},
-  author = {Tyson Barrett and Matt Dowle and Arun Srinivasan and Jan Gorecki and Michael Chirico and Toby Hocking},
+  author = {Barrett, Tyson and Dowle, Matt and Srinivasan, Arun and Gorecki, Jan and Chirico, Michael and Hocking, Toby},
   note   = {R package version 1.15.0},
   year   = {2024},
   url    = {https://CRAN.R-project.org/package=data.table},
@@ -158,7 +158,7 @@
 
 @Manual{Ooms2024,
   title  = {{magick}: Advanced Graphics and Image-Processing in {R}},
-  author = {Jeroen Ooms},
+  author = {Ooms, Jeroen},
   note   = {R package version 2.8.3},
   year   = {2024},
[…]
