### 2024-05-21 9790d67
Fix spelling and citation issues
