+++ inst/paper/paper.md
@@ -33,8 +33,8 @@ authors:
   orcid: 0000-0002-0061-8359
 date: "25 Mar 2024"
 output:
-  html_document: default
   pdf_document: default
+  html_document: default
 bibliography: paper.bib
 tags:
 - Australia
@@ -65,14 +65,13 @@ affiliations:
   name: University of Southern Queensland, Centre for Crop Health, Toowoomba, Queensland 4350, Australia
 - index: 9
   name: Curtin University, Centre for Crop and Disease Management, Curtin Biometry and Agricultural Data Analytics, Bentley, Western Australia 6102, Australia
-
 ---
 
 # Summary
 
-Researchers and policymakers use weather data in a variety of ways. 
+Researchers and policy makers use weather data in a variety of ways.
 Agriculture applications of the data are used in several types of models and decision support tools, to estimate leaf wetness, crop yield, crop growth stage, physiological stress or forecasting crop disease epidemics or insect pest population levels [@Venaelaeinen2002; @DeWolf2003; Sparks2017].
-Other areas of use include mapping potential renewable energy, _e.g._, wind or solar potential for exploration purposes [@Ramachandra2007].
+Other areas of use include mapping potential renewable energy, *e.g.*, wind or solar potential for exploration purposes [@Ramachandra2007].
 The data can also be used by decision makers for municipalities to help plan for extreme weather events, energy needs and other infrastructure [@Svensson2002; @Alcoforado2009].
 
 {weatherOz} offers R users with a single interface to access Australian climate and weather data sources providing a standardised way of easily querying and retrieving Australian climate and weather data.
@@ -80,26 +79,27 @@ The data can also be used by decision makers for municipalities to help plan for
 # Statement of need
 
 Australian weather data availability is fragmented and difficult to easily access repeatedly or programmatically for use in research activities.
-Previously the R package, {bomrang} [@Sparks2017], provided easy access to data available from the Australian Bureau of Meteorology (BoM) in R.
+Previously the R package, {bomrang} [@Sparks2017a], provided easy access to data available from the Australian Bureau of Meteorology (BoM) in R.
 However, difficulties with accessing the data and restrictive polices caused us to archive this package.
 Therefore, we created {weatherOz} to replace and at the same time, improve upon {bomrang} and provide R users with a package that allowed for easy of access to multiple sources of Australian climate and weather data while providing a unified approach to working with different data sources.
-There are two sources of Australia-wide climate and weather station observations and data, BoM and Queensland Government's Scientific Information for Landowners (SILO) [@Jeffery2001] database, hosted by the Queensland Department of the Environment, Science and Innovation (DESI) and one source for Western Australia only weather station data, the Department of Primary Industries and Regional Development of Western Australia's (DPIRD) database.
-BoM provides files through their public anonymous FTP server [@BoM2024] and data from both the SILO and DPIRD databases are available through APIs.
+There are two sources of Australia-wide climate and weather station observations and data, BoM and Queensland Government's Scientific Information for Landowners (SILO) [@Jeffrey2001] database, hosted by the Queensland Department of the Environment, Science and Innovation (DESI) and one source for Western Australia only weather station data, the Department of Primary Industries and Regional Development of Western Australia's (DPIRD) database.
+BoM provides files through their public anonymous FTP server [@BOM2024] and data from both the SILO and DPIRD databases are available through APIs.
 
 {weatherOz} provides access to data from the SILO database are made available under a Creative Commons Attribution 4.0 International (CC BY 4.0) licence including the Patched Point data available "from 1889 to yesterday" for approximately 8,000 weather station locations located at post-offices, airports, police stations, national parks and private properties [@SILO2024], and Data Drill data, which are spatially interpolated data covering Australian land surfaces.
[…]
