### 2024-05-21 1d5b5f7
Ensure all SILO stations are present even not if in BOM data
