### 2024-05-21 29bc3e6
Fix station names in SILO stations metadata
