+++ R/get_stations_metadata.R
@@ -226,7 +226,8 @@ get_stations_metadata <-
       .dataset = "PatchedPoint"
   )
 
-  station_metadata <- bom_stations[silo_stations, on = "station_code"]
+  station_metadata <- bom_stations[silo_stations, on = c("station_code",
+                                                         "station_name")]
   # drops the unwanted columns that are added after using `find_nearby_stations`
   station_metadata[, grep("^i.", names(station_metadata)) := NULL]
   station_metadata[, owner := NULL]
