### 2024-05-21 6bca894
Remove redundant code
