### 2024-05-21 68cbde6
fix lon/lat check
