### 2024-05-21 2d41e85
return an sf_polygon not matrix from OSM query
