+++ tests/testthat/test-find_stations_in.R
@@ -94,7 +94,7 @@ test_that("find_stations_in() works with bbox", {
 test_that("find_stations_in() works with place name", {
   vcr::use_cassette("find_stations_in_place_name", {
     x <- find_stations_in(x = "Toowoomba Qld",
-                          which_api = "SILO",
+                          which_api = "silo",
                           include_closed = TRUE)
   })
   expect_length(x, 11)
