### 2024-05-21 4c8a333
Add Max to authors' list
