+++ DESCRIPTION
@@ -22,7 +22,9 @@ Authors@R: c(
     person("Keith", "Pembleton", , "keith.pembleton@usq.edu.au", role = "aut",
            comment = c(ORCID = "0000-0002-1896-4516")),
     person("Maëlle", "Salmon", , "maelle.salmon@yahoo.se", role = "ctb",
-           comment = c(ORCID = "0000-0002-2815-0399", "Contributed to debugging a nasty little bug with CI where timezones caused tests to fail due to vcr not recognising the\n                URL when run outside of Australia/Perth TZ! Suggested the use of `withr::local_timzeone()`, see <https://github.com/ropensci/weatherOz/commit/b052bf91973b8d7e147a39e8938405a64622634b>.")),
+           comment = c(ORCID = "0000-0002-2815-0399", "Contributed to debugging a nasty little bug with CI where timezones caused tests to fail due to 'vcr' not recognising the\n                URL when run outside of Australia/Perth TZ! Suggested the use of `withr::local_timzeone()`, see <https://github.com/ropensci/weatherOz/commit/b052bf91973b8d7e147a39e8938405a64622634b>.")),
+    person("Max", "Moldovan", , "max.moldovan@adelaide.edu.au", role = "ctb",
+           comment = c(ORCID = "0000-0001-9680-8474", "Contributed valuable feedback on package usage leading to improvements in the package structure and functionality.")),
     person("Jimmy", "Ng", , "jimmy.ng@dpird.wa.gov.au", role = "ctb"),
     person("Steve", "Collins", , "steve.collins@dpird.wa.gov.au", role = "ctb",
            comment = "Designed weatherOz's hex logo."),
