### 2024-05-21 af3ad10
Ensure no missing lat/lon values in metadata
