### 2024-05-21 efa301a
remove knitted vignettes and update wordlist
