+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -1,1297 +1,0 @@
-
----
-title: "weatherOz for DPIRD"
-author: "Rodrigo Pires, Anna Hepworth, Rebecca O'Leary and Adam H. Sparks"
-output:
-  rmarkdown::html_vignette:
-    toc: true
-vignette: >
-  %\VignetteIndexEntry{weatherOz for DPIRD}
-  %\VignetteEngine{knitr::rmarkdown_notangle}
-  %\VignetteEncoding{UTF-8}
----
-
-
-
-# Working With DPIRD Data
-
-Three functions are provided to streamline fetching data from the DPIRD Weather 2.0 API endpoints.
-
- * `get_dpird_extremes()`, which returns the recorded extreme values for the given station in the DPIRD weather station network.;
- * `get_dpird_minute()`, which returns weather data in minute increments for stations in the DPIRD weather station network with only the past two years being available; and
- * `get_dpird_summaries()`, which returns weather data in 15 and 30 minute, hourly, daily, monthly or yearly summary values for stations in the DPIRD weather station network.
-
-## A Note on API Keys
-
-All examples in this vignette assume that you have stored your API key in your .Renviron file.
-See [Chapter 8](https://rstats.wtf/r-startup.html#renviron) in "What They Forgot to Teach You About R" by Bryan _et al._ for more on storing details in your .Renviron if you are unfamiliar.
-
-## Getting Extreme Weather Values
-
-The `get_dpird_extremes()` function fetches and returns nicely formatted individual extreme weather summaries from the DPIRD Weather 2.0 API.
-You must provide a `station_code` and `API_key`, the other arguments, `values` and `include_closed` are optional.
-
-### Available Values for Extreme Weather
-
- * all (which will return all of the following values),
- * erosionCondition,
- * erosionConditionLast7Days,
- * erosionConditionLast7DaysDays,
- * erosionConditionLast7DaysMinutes,
- * erosionConditionLast14Days,
- * erosionConditionLast14DaysDays,
- * erosionConditionLast14DaysMinutes,
- * erosionConditionMonthToDate,
- * erosionConditionMonthToDateDays,
- * erosionConditionMonthToDateMinutes,
- * erosionConditionMonthToDateStartTime,
- * erosionConditionSince12AM,
- * erosionConditionSince12AMMinutes,
- * erosionConditionSince12AMStartTime,
- * erosionConditionYearToDate,
- * erosionConditionYearToDateDays,
- * erosionConditionYearToDateMinutes,
- * erosionConditionYearToDateStartTime,
- * frostCondition,
- * frostConditionLast7Days,
- * frostConditionLast7DaysDays,
- * frostConditionLast7DaysMinutes,
- * frostConditionLast14Days,
- * frostConditionLast14DaysDays,
- * frostConditionLast14DaysMinutes,
- * frostConditionMonthToDate,
- * frostConditionMonthToDateDays,
- * frostConditionMonthToDateMinutes,
- * frostConditionMonthToDateStartTime,
- * frostConditionSince9AM,
- * frostConditionSince9AMMinutes,
- * frostConditionSince9AMStartTime,
- * frostConditionTo9AM,
- * frostConditionTo9AMMinutes,
- * frostConditionTo9AMStartTime,
- * frostConditionYearToDate,
- * frostConditionYearToDate,
- * frostConditionYearToDateMinutes,
- * frostConditionYearToDateStartTime,
- * heatCondition,
- * heatConditionLast7Days,
- * heatConditionLast7DaysDays,
- * heatConditionLast7DaysMinutes,
- * heatConditionLast14Days,
- * heatConditionLast14DaysDays,
[…]
