+++ vignettes/weatherOz_for_BOM.Rmd
@@ -1,738 +1,0 @@
-
----
-title: "weatherOz for BOM"
-author: "Jonathan Carroll, Dean Marchiori, Paul Melloy, Mark Padgham, Hugh Parsonage, Keith Pembleton and Adam H. Sparks"
-output:
-  rmarkdown::html_vignette:
-    toc: true
-vignette: >
-  %\VignetteIndexEntry{weatherOz for BOM}
-  %\VignetteEngine{knitr::rmarkdown_notangle}
-  %\VignetteEncoding{UTF-8}
----
-
-
-# Using {weatherOz} With BOM Data
-
-Several functions are provided by {weatherOz} to retrieve Australian Bureau of Meteorology (BOM) data.
-A family of functions retrieve data files from BOM or parse local files from BOM and return data frames;
-
- * `get_precis_forecast()` and `parse_precis_forecast(), which return the précis (short) forecast;
- * `get_ag_bulletin()` and `parse_ag_bulletin(), which return the agriculture bulletin;
- * `get_coastal_forecast()` and `parse_coastal_forecast(), which return coastal waters forecasts for each state.
-
-## Using get_precis_forecast()
-
-This function only takes one argument, `state`. The `state` parameter allows the user to select the forecast for just one state or a national forecast.
-States or territories are specified using the official postal codes or full name with fuzzy matching performed via `agrep()`.
-
-- **ACT** - Australian Capital Territory
-
-- **NSW** - New South Wales
-
-- **NT** - Northern Territory
-
-- **QLD** - Queensland
-
-- **SA** - South Australia
-
-- **TAS** - Tasmania
-
-- **VIC** - Victoria
-
-- **WA** - Western Australia
-
-- **AUS** - Australia, returns national forecast including all states, NT and ACT.
-
-### Results
-
-The function, `get_precis_forecast()`, will return a data frame of the weather forecast for the daily forecast for selected towns.
-See Appendix 1 for a full description of the fields and values.
-
-### Example 1: Getting a Forecast
-
-Following is an example fetching the forecast for Queensland.
-
-
-``` r
-library(weatherOz)
-(QLD_forecast <- get_precis_forecast(state = "QLD"))
-#>   -----------  Australian Bureau of Meteorology (BOM) Precis Forecast. -
-#>   The HTML version of Short Form (Precis) Forecast for 
-#>   QLD can be found at:
-#>   <http://www.bom.gov.au/qld/forecasts/state.shtml>
-#>   Please note information at the page 
-#>   <http://www.bom.gov.au/catalogue/data-feeds.shtml#precis>.
-#>   --------------------------------------------------------------------------------  
-#>       index product_id  state         town       aac       lat      lon  elev
-#>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>
-#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1
-#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1
-#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1
-#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1
-#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1
-#>  ---                                                                         
-#> 900:      3   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4
[…]

[…]
