### 2024-05-21 8e1b580
Remove {osmdata} API named-place approach, API is too flakey
