### 2024-05-21 280e4f0
Reknit vignettes on Raspberry Pi
