+++ vignettes/weatherOz.Rmd
@@ -1,0 +1,289 @@
+---
+title: "weatherOz"
+author: "Rodrigo Pires, Anna Hepworth, Rebecca O'Leary, Jonathan Carroll, Dean Marchiori, Paul Melloy, Mark Padgham, Hugh Parsonage, Keith Pembleton and Adam H. Sparks"
+output:
+  rmarkdown::html_vignette:
+    toc: true
+vignette: >
+  %\VignetteIndexEntry{weatherOz}
+  %\VignetteEngine{knitr::rmarkdown_notangle}
+  %\VignetteDepends{ggplot2}
+  %\VignetteDepends{ggthemes}
+  %\VignetteDepends{grid}
+  %\VignetteDepends{gridextra}
+  %\VignetteDepends{mapproj}
+  %\VignetteDepends{maps}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+## Introduction
+
+{weatherOz} provides automated downloading, parsing, and formatting of weather data for Australia through API endpoints provided by the Department of Primary Industries and Regional Development (DPIRD) of Western Australia, and by the Science and Technology Division of the Queensland Government's Department of Environment and Science (DES).
+As well as Australian Government Bureau of Meteorology ('BOM') précis and coastal forecasts, agriculture bulletin data, and downloading and importing radar and satellite imagery files.
+
+DPIRD weather data are accessed through public APIs provided by DPIRD, <https://www.agric.wa.gov.au/weather-api-20>, providing access to weather station data from DPIRD's own weather station network.
+Detailed information on using {weatherOz} with DPIRD data is available in the [weatherOz for the DPIRD Weather 2.0 API](weatherOz_for_DPIRD.html) vignette.
+
+Australia-wide weather data are based on data from the Australian Bureau of Meteorology (BOM) data and accessed through SILO (Scientific Information for Land Owners) (Jeffery et al., 2001).
+More details on using {weatherOz} with SILO data are available in the [weatherOz for the SILO API](weatherOz_for_SILO.html) vignette.
+
+The BOM also serves several types of data data as XML, JSON and SHTML files. This package fetches these files, parses them and return a data frame.
+Satellite and radar imagery files are also made available to the public via anonymous FTP.
+{weatherOz} provides functionality to query, fetch and create `terra::SpatRaster()` or {stars} objects of GeoTIFF satellite imagery or a {magick} object of radar image.png files.
+For detailed information on using {weatherOz} with BOM data, please see the [weatherOz for BOM Data Resources](weatherOz_for_BOM.html) vignette.
+
+Following is a quick demonstration of some common tasks that you may wish to undertake while using {weatherOz}.
+
+## A Note on API Keys
+
+All examples in this vignette assume that you have stored your API key in your .Renviron file.
+See [Chapter 8](https://rstats.wtf/r-startup.html#renviron) in "What They Forgot to Teach You About R" by Bryan _et al._ for more on storing details in your .Renviron if you are unfamiliar.
+
+## Using {weatherOz} to Fetch Daily Summary DPIRD Data
+
+Daily summary weather data is frequently used.
+The following example will demonstrate how you can find and download weather station data for stations in or near Northam, WA for the year of 2022.
+This example assumes that you have stored your DPIRD API key in your .Renviron file.
+
+
+``` r
+library(weatherOz)
+#> 
+#> Attaching package: 'weatherOz'
+#> The following object is masked from 'package:graphics':
+#> 
+#>     plot
+#> The following object is masked from 'package:base':
+#> 
+#>     plot
+```
+
+``` r
+
[…]
