+++ vignettes/weatherOz_for_SILO.Rmd
@@ -1,0 +1,797 @@
+
+---
+title: "weatherOz for SILO"
+author: "Rodrigo Pires, Anna Hepworth, Rebecca O'Leary and Adam H. Sparks"
+output:
+  rmarkdown::html_vignette:
+    toc: true
+vignette: >
+  %\VignetteIndexEntry{weatherOz for SILO}
+  %\VignetteEngine{knitr::rmarkdown_notangle}
+  %\VignetteEncoding{UTF-8}
+bibliography: bibliography.bib
+---
+
+
+
+# About SILO Data
+
+[Queensland Scientific Information for Landowners (SILO)](https://www.longpaddock.qld.gov.au/silo/about/) data sets are constructed from observational records provided by the Bureau of Meteorology (BOM).
+SILO interpolates the raw data to fill missing data creating data sets that are spatially and temporally complete [@Jeffrey2001].
+
+# Working With SILO Data
+
+Four functions are provided to streamline fetching data from the SILO API endpoints.
+
+ * `get_data_drill()`, which returns weather data from SILO's 0.05 x 0.05 degrees interpolated, gridded weather dataset;
+ * `get_patched_point()`, which returns weather data from a given BOM weather station (a point) in the SILO database where missing values are interpolated (patched);
+ * `get_data_drill_apsim()`, which returns data drill weather data in an {apsimx} `.met` class object with the weather data ready for use in APSIM; and
+ * `get_patched_point_apsim()`, which returns patched point weather data in an {apsimx} `.met` class object with the weather data ready for use in APSIM.
+
+## Available Values
+
+`get_data_drill()` and `get_patched_point()` both accept the following values.
+`get_data_drill_apsim()` and `get_patched_point_apsim()` return a canned set of values for use in APSIM, because of this, they do not accept a `values` argument.
+
+<dl>
+  <dt>all</dt><dd>Which will return all of the following values</dd>
+  <dt>max_temp (degrees C)</dt><dd>Maximum temperature</dd>
+  <dt>min_temp (degrees C)</dt><dd>Minimum temperature</dd>
+  <dt>vp (hPa)</dt><dd>Vapour pressure</dd>
+  <dt>vp_deficit (hPa)</dt><dd>Vapour pressure deficit</dd>
+  <dt>evap_pan (mm)</dt><dd>Class A pan evaporation</dd>
+  <dt>evap_syn (mm)</dt><dd>Synthetic estimate [@Rayner2005]</dd>
+  <dt>evap_comb (mm)</dt><dd>Combination (synthetic estimate pre-1970, class A pan 1970 onwards)</dd>
+  <dt>evap_morton_lake (mm)</dt><dd>Morton's shallow lake evaporation</dd>
+  <dt>radiation (Mj/m [@Morton1983])</dt><dd>Solar exposure, consisting of both direct and diffuse components</dd>
+  <dt>rh_tmax (%)</dt><dd>Relative humidity at the time of maximum temperature</dd>
+  <dt>rh_tmin (%)</dt><dd>Relative humidity at the time of minimum temperature</dd>
+  <dt>et_short_crop (mm)</dt><dd>FAO56 [@Allen1998] short crop</dd>
+  <dt>et_tall_crop (mm)</dt><dd>ASCE [@Walter2000] tall crop [@Jeffrey2001]</dd>
+  <dt>et_morton_actual (mm)</dt><dd>Morton's areal actual evapotranspiration</dd>
+  <dt>et_morton_potential (mm)</dt><dd>Morton's point potential evapotranspiration</dd>
+  <dt>et_morton_wet (mm)</dt><dd>Morton's wet-environment areal potential evapotranspiration over land</dd>
+  <dt>mslp (hPa)</dt><dd>Mean sea level pressure</dd>
+</dl>
+
+## Value information
+
[…]

[…]
