### 2024-05-21 cd65f3e
Remove tests for place name
