### 2024-05-21 0b883ec
Clean up documentation and check user-inputs
