### 2024-05-21 8ae2d24
remove check for input x
