### 2024-05-22 0cf0215
Fix error message for users
