### 2024-05-26 ee68fe4
(Re)Add use case vignette that was somehow missing
