+++ vignettes/use_case.Rmd
@@ -1,0 +1,155 @@
+---
+title: "weatherOz Use Case"
+author: "Keith Pembleton, Adam H. Sparks and Mark Padgham"
+output:
+  rmarkdown::html_vignette:
+    toc: true
+vignette: >
+  %\VignetteIndexEntry{weatherOz Use Case}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+  %\VignetteDepends{dplyr}
+  %\VignetteDepends{mailR}
+bibliography: bibliography.bib
+---
+
+
+
+## Introduction
+
+[Summer pulses (*e.g.*, mungbeans) are a high value crop for Queensland grain growers](https://www.pulseaus.com.au/about/australian-pulse-industry).
+However, temperatures above 40 °C, particularly during flowering, can have a severe detrimental effect on crop yields [@Kaushal2016].
+A key strategy that plants use to minimise the impact of heat stress is evaporative cooling.
+Ensuring crops are well watered prior to a heat stress event (so to facilitate high rates of evaporative cooling) is a potential strategy to ensure that irrigated crops avoid heat stress.
+However, this requires farmers to recognise a heat stress event is coming and commence irrigating with enough time to water the entire crop prior to the event occurring.
+
+The [{WINS}](https://github.com/ToowoombaTrio/WINS) for pulses project provides warnings when heat stress events are likely to occur so that farmers can irrigate their crops prior to the event to minimise any resulting yield
+loss.
+It does this by using the `get_precis_forecast()` function from {weatherOz} along with functionality from
+[{mailR}](https://CRAN.R-project.org/package=mailR) [@Premraj2021].
+
+## Materials and Methods
+
+
+``` r
+install.packages("weatherOz")
+install.packages("mailR")
+library(dplyr)  # for filter()
+library(weatherOz)
+library(mailR)
+```
+
+For this to work we require a list of alert subscribers.
+Here we create a dummy set of subscribers, locations, and email addresses.
+
+
+``` r
+subscribers_list <-
+  data.frame(cbind(
+    c(1, 2, 3),
+    c("Joe", "John", "Jayne"),
+    c("Blogs", "Doe", "Doe"),
+    c("Dalby", "Toowoomba", "Warwick"),
+    c(
+      "XXXX.XXXX@gmail.com",
+      "XXXX.XXXX@hotmail.com",
+      "123.123@gmail.com"
+    )
+  ))
+colnames(subscribers_list) <-
+  c("Entry", "Name", "Surname", "Location", "email")
+head(subscribers_list)
+```
+
+```
+##   Entry  Name Surname  Location                 email
+## 1     1   Joe   Blogs     Dalby   XXXX.XXXX@gmail.com
+## 2     2  John     Doe Toowoomba XXXX.XXXX@hotmail.com
+## 3     3 Jayne     Doe   Warwick     123.123@gmail.com
+```
+
+We also need to specify our email address and password.
+For this project we are using a gmail account to send the email alerts.
+
+
+``` r
+our_email <- "yyyy***xxxx@gmail.com"
+our_password <- "*password*"
+```
+
+Finally we need to set a forecast temperature threshold that will trigger an alert to be sent.
+
+
+``` r
+threshold_temp <- 40
+```
+
+We then use {weatherOz} to access the latest précis forecast for Queensland.
+
+
+``` r
+QLD_forecast <- get_precis_forecast(state = "QLD")
+```
+
+Now we use a `for()` loop that searches through the Queensland précis forecast data frame and subsets it based on location each subscriber has nominated and the temperature threshold we have set.
+As part of the `for()` loop there is an `if()` statement that uses the {mailR} package to send an email to the
[…]
