+++ vignettes/weatherOz_for_BOM.Rmd
@@ -66,30 +66,30 @@ library(weatherOz)
 #>   --------------------------------------------------------------------------------  
 #>       index product_id  state         town       aac       lat      lon  elev    start_time_local end_time_local utc_offset      start_time_utc        end_time_utc minimum_temperature maximum_temperature lower_precipitation_limit upper_precipitation_limit
 #>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>              <POSc>         <POSc>     <fctr>              <POSc>              <POSc>               <num>               <num>                     <num>                     <num>
-#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-21 17:00:00     2024-05-22      10:00 2024-05-21 07:00:00 2024-05-21 14:00:00                  NA                  NA                        NA                        NA
-#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-22 00:00:00     2024-05-23      10:00 2024-05-21 14:00:00 2024-05-22 14:00:00                  13                  24                        NA                        NA
-#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-23 00:00:00     2024-05-24      10:00 2024-05-22 14:00:00 2024-05-23 14:00:00                  12                  24                        NA                        NA
-#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-24 00:00:00     2024-05-25      10:00 2024-05-23 14:00:00 2024-05-24 14:00:00                  12                  25                        NA                        NA
-#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-25 00:00:00     2024-05-26      10:00 2024-05-24 14:00:00 2024-05-25 14:00:00                  14                  23                         0                         1
+#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-26 09:07:35     2024-05-27      10:00 2024-05-25 23:07:35 2024-05-26 14:00:00                  NA                  24                         0                         1
+#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-27 00:00:00     2024-05-28      10:00 2024-05-26 14:00:00 2024-05-27 14:00:00                  13                  25                        NA                        NA
+#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-28 00:00:00     2024-05-29      10:00 2024-05-27 14:00:00 2024-05-28 14:00:00                  13                  24                         0                         1
+#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-29 00:00:00     2024-05-30      10:00 2024-05-28 14:00:00 2024-05-29 14:00:00                  14                  25                         0                         1
+#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-30 00:00:00     2024-05-31      10:00 2024-05-29 14:00:00 2024-05-30 14:00:00                  15                  24                         0                         2
 #>  ---                                                                                                                                                                                                                                                           
[…]
