+++ vignettes/weatherOz_for_SILO.Rmd
@@ -14,12 +14,17 @@ bibliography: bibliography.bib
 
 
 
-# About SILO Data
+## About SILO Data
 
 [Queensland Scientific Information for Landowners (SILO)](https://www.longpaddock.qld.gov.au/silo/about/) data sets are constructed from observational records provided by the Bureau of Meteorology (BOM).
 SILO interpolates the raw data to fill missing data creating data sets that are spatially and temporally complete [@Jeffrey2001].
 
-# Working With SILO Data
+## A Note on API Keys
+
+All examples in this vignette assume that you have stored your API key in your .Renviron file.
+See [Chapter 8](https://rstats.wtf/r-startup.html#renviron) in "What They Forgot to Teach You About R" by Bryan _et al._ for more on storing details in your .Renviron if you are unfamiliar.
+
+## Working With SILO Data
 
 Four functions are provided to streamline fetching data from the SILO API endpoints.
 
@@ -54,7 +59,7 @@ Four functions are provided to streamline fetching data from the SILO API endpoi
   <dt>mslp (hPa)</dt><dd>Mean sea level pressure</dd>
 </dl>
 
-## Value information
+## Value Information
 
 Solar radiation: total incoming downward shortwave radiation on a horizontal surface, derived from estimates of cloud oktas and sunshine duration [@Zajaczkowski2013].
 
@@ -76,11 +81,6 @@ Where possible (depending on the file format), the data are supplied with codes 
 </dl>
 
 The data returned will include columns that denote the source of the data, whether it was an observation or interpolated.
-
-## A Note on API Keys
-
-All examples in this vignette assume that you have stored your API key in your .Renviron file.
-See [Chapter 8](https://rstats.wtf/r-startup.html#renviron) in "What They Forgot to Teach You About R" by Bryan _et al._ for more on storing details in your .Renviron if you are unfamiliar.
 
 ## Getting Data Drill Interpolated Data
 
@@ -113,24 +113,24 @@ library(weatherOz)
 #>   4:    150.05   -27.85  2022     1     4 2022-01-04     36.3              25     18.1              25 254.5 m              5.1                      26                11.1                         26           8.1                   26           7.7                   26
 #>   5:    150.05   -27.85  2022     1     5 2022-01-05     36.4              25     20.3              25 254.5 m              5.0                      26                10.9                         26           8.0                   26           7.5                   26
 #>  ---                                                                                                                                                                                                                                                                        
-#> 867:    150.05   -27.85  2024     5    16 2024-05-16     24.5              25     11.8              25 254.5 m              1.2                      26                 3.3                         26           2.2                   26           2.4                   26
-#> 868:    150.05   -27.85  2024     5    17 2024-05-17     23.7              25     12.9              25 254.5 m              2.1                      26                 3.2                         26           2.6                   26           2.3                   26
-#> 869:    150.05   -27.85  2024     5    18 2024-05-18     21.8              25      8.8              25 254.5 m              1.7                      26                 3.6                         26           2.6                   26           2.4                   26
-#> 870:    150.05   -27.85  2024     5    19 2024-05-19     20.2              25      5.8              25 254.5 m              0.8                      26                 4.0                         26           2.4                   26           2.5                   26
-#> 871:    150.05   -27.85  2024     5    20 2024-05-20     22.2              75      3.0              25 254.5 m              0.7                      26                 4.1                         26           2.4                   26           2.7                   26
[…]

[…]
