### 2024-05-26 1534d68
Fixes lon/lat being character values in BOM/SILO metadata
