### 2024-05-26 cc9588c
Reknit vignettes with corrected BOM metadata
