+++ vignettes/weatherOz.Rmd
@@ -153,17 +153,17 @@ ppd <- get_patched_point(
 ppd
 #>      station_code station_name  year month   day       date air_tmax air_tmax_source air_tmin air_tmin_source  elev_m  extracted latitude longitude rainfall rainfall_source rh_tmax rh_tmax_source rh_tmin rh_tmin_source
 #>            <fctr>       <char> <num> <num> <int>     <Date>    <num>           <int>    <num>           <int>  <char>     <Date>    <num>     <num>    <num>           <int>   <num>          <int>   <num>          <int>
-#>   1:       041103    Toowoomba  2022     1     1 2022-01-01     22.8              25     16.6              25 691.0 m 2024-05-21 -27.5836  151.9317      7.6              25    72.1             26   100.0             26
-#>   2:       041103    Toowoomba  2022     1     2 2022-01-02     27.7              25     14.2              25 691.0 m 2024-05-21 -27.5836  151.9317      0.1              25    47.1             26   100.0             26
-#>   3:       041103    Toowoomba  2022     1     3 2022-01-03     29.4              25     13.8              25 691.0 m 2024-05-21 -27.5836  151.9317      0.0              25    38.8             26   100.0             26
-#>   4:       041103    Toowoomba  2022     1     4 2022-01-04     30.0              25     16.8              25 691.0 m 2024-05-21 -27.5836  151.9317      0.0              25    43.1             26    95.7             26
-#>   5:       041103    Toowoomba  2022     1     5 2022-01-05     29.5              25     17.5              25 691.0 m 2024-05-21 -27.5836  151.9317      0.0              25    49.2             26   100.0             26
+#>   1:       041103    Toowoomba  2022     1     1 2022-01-01     22.8              25     16.6              25 691.0 m 2024-05-26 -27.5836  151.9317      7.6              25    72.1             26   100.0             26
+#>   2:       041103    Toowoomba  2022     1     2 2022-01-02     27.7              25     14.2              25 691.0 m 2024-05-26 -27.5836  151.9317      0.1              25    47.1             26   100.0             26
+#>   3:       041103    Toowoomba  2022     1     3 2022-01-03     29.4              25     13.8              25 691.0 m 2024-05-26 -27.5836  151.9317      0.0              25    38.8             26   100.0             26
+#>   4:       041103    Toowoomba  2022     1     4 2022-01-04     30.0              25     16.8              25 691.0 m 2024-05-26 -27.5836  151.9317      0.0              25    43.1             26    95.7             26
+#>   5:       041103    Toowoomba  2022     1     5 2022-01-05     29.5              25     17.5              25 691.0 m 2024-05-26 -27.5836  151.9317      0.0              25    49.2             26   100.0             26
 #>  ---                                                                                                                                                                                                                      
-#> 361:       041103    Toowoomba  2022    12    27 2022-12-27     26.2              25     15.1              25 691.0 m 2024-05-21 -27.5836  151.9317      0.0              25    47.9             26    95.0             26
-#> 362:       041103    Toowoomba  2022    12    28 2022-12-28     26.7              25     13.0              25 691.0 m 2024-05-21 -27.5836  151.9317      0.2              25    43.7             26   100.0             26
-#> 363:       041103    Toowoomba  2022    12    29 2022-12-29     27.6              25     14.1              25 691.0 m 2024-05-21 -27.5836  151.9317      0.0              25    44.1             26   100.0             26
-#> 364:       041103    Toowoomba  2022    12    30 2022-12-30     27.0              25     17.0              25 691.0 m 2024-05-21 -27.5836  151.9317      0.7              25    60.3             26   100.0             26
[…]
