+++ vignettes/weatherOz_for_BOM.Rmd
@@ -66,7 +66,7 @@ library(weatherOz)
 #>   --------------------------------------------------------------------------------  
 #>       index product_id  state         town       aac       lat      lon  elev    start_time_local end_time_local utc_offset      start_time_utc        end_time_utc minimum_temperature maximum_temperature lower_precipitation_limit upper_precipitation_limit
 #>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>              <POSc>         <POSc>     <fctr>              <POSc>              <POSc>               <num>               <num>                     <num>                     <num>
-#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-26 09:07:35     2024-05-27      10:00 2024-05-25 23:07:35 2024-05-26 14:00:00                  NA                  24                         0                         1
+#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-26 13:34:34     2024-05-27      10:00 2024-05-26 03:34:34 2024-05-26 14:00:00                  NA                  24                         0                         1
 #>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-27 00:00:00     2024-05-28      10:00 2024-05-26 14:00:00 2024-05-27 14:00:00                  13                  25                        NA                        NA
 #>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-28 00:00:00     2024-05-29      10:00 2024-05-27 14:00:00 2024-05-28 14:00:00                  13                  24                         0                         1
 #>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-05-29 00:00:00     2024-05-30      10:00 2024-05-28 14:00:00 2024-05-29 14:00:00                  14                  25                         0                         1
@@ -77,19 +77,19 @@ library(weatherOz)
 #> 789:      4   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-05-30 00:00:00     2024-05-31      10:00 2024-05-29 14:00:00 2024-05-30 14:00:00                  21                  26                         0                         3
 #> 790:      5   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-05-31 00:00:00     2024-06-01      10:00 2024-05-30 14:00:00 2024-05-31 14:00:00                  21                  27                         0                         1
 #> 791:      6   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-01 00:00:00     2024-06-02      10:00 2024-05-31 14:00:00 2024-06-01 14:00:00                  20                  27                         0                         1
-#>                          precis probability_of_precipitation
-#>                          <char>                        <num>
-#>   1: Possible afternoon shower.                           40
[…]
