### 2024-05-26 2c34753
Closes https://github.com/ropensci/weatherOz/issues/73
