### 2024-05-26 21f1fd2
Closes https://github.com/ropensci/weatherOz/issues/74
