+++ inst/paper/paper.bib
@@ -126,6 +126,7 @@
   author = {Sparks, Adam and Pembleton, Keith and Grundy, Gordon},
   note   = {R package version 0.0.1.9000},
   year   = {2017},
+  doi    = {10.5281/zenodo.11363275},
   url    = {https://github.com/ToowoombaTrio/WINS},
 }
 
