### 2024-05-28 3f0e628
fixes yet one more issue with references
