+++ R/globals.R
@@ -21,6 +21,8 @@ utils::globalVariables(c(
   "station_name", # <.get_bom_metadata>
   "start", # <.get_bom_metadata>
   "end", # <.get_bom_metadata>
+  "latitude", # <.get_bom_metadata>
+  "longitude", # <.get_bom_metadata>
   "status", # <.get_bom_metadata>
   "elev_m", # <.get_bom_metadata>
   "wmo", # <.get_bom_metadata>
