### 2024-05-29 9b76562
Adds functionality to find stations in a geographic area (#71)
