+++ R/dpird_extreme_weather_values.R
@@ -1,17 +1,0 @@
-
-#' A List of DPIRD Extreme Weather Data Values
-#'
-#' A [vector] object containing 57 items representing valid values to supply
-#'   to `get_dpird_extremes()`'s \var{values} argument taken from the
-#'   documentation for the \acronym{DPIRD} Weather 2.0 \acronym{API}.
-#'
-#' @format A [vector] object of 57 items.
-#'
-#' @keywords datasets
-#'
-#' @family DPIRD
-#' @family data
-#'
-#' @source <https://www.agric.wa.gov.au/weather-api-20>
-#'
-"dpird_extreme_weather_values"
