### 2024-05-29 d754d62
try to fix pkgdown YAML
