### 2024-05-29 eec81d3
Correct documentation for `find_stations_in()`
