### 2024-06-02 #6717 @arfon
@bozaah – this all looks good now. The only issue I can see now is that all of the software names that are wrapped in `{}` should be updated to be simply wrapped in backticks. e.g.:

```
Change from this:

`{weatherOz}`

...to this:

`weatherOz`
```
