+++ inst/paper/paper.md
@@ -74,18 +74,18 @@ Agriculture applications of the data are used in several types of models and dec
 Other areas of use include mapping potential renewable energy, *e.g.*, wind or solar potential for exploration purposes [@Ramachandra2007].
 The data can also be used by decision makers for municipalities to help plan for extreme weather events, energy needs and other infrastructure [@Svensson2002; @Alcoforado2009].
 
-{weatherOz} offers R users with a single interface to access Australian climate and weather data sources providing a standardised way of easily querying and retrieving Australian climate and weather data.
+`weatherOz` offers R users with a single interface to access Australian climate and weather data sources providing a standardised way of easily querying and retrieving Australian climate and weather data.
 
 # Statement of need
 
 Australian weather data availability is fragmented and difficult to easily access repeatedly or programmatically for use in research activities.
-Previously the R package, {bomrang} [@Sparks2017a], provided easy access to data available from the Australian Bureau of Meteorology (BoM) in R.
+Previously the R package, `bomrang` [@Sparks2017a], provided easy access to data available from the Australian Bureau of Meteorology (BoM) in R.
 However, difficulties with accessing the data and restrictive polices caused us to archive this package.
-Therefore, we created {weatherOz} to replace and at the same time, improve upon {bomrang} and provide R users with a package that allowed for easy of access to multiple sources of Australian climate and weather data while providing a unified approach to working with different data sources.
+Therefore, we created `weatherOz` to replace and at the same time, improve upon `bomrang` and provide R users with a package that allowed for easy of access to multiple sources of Australian climate and weather data while providing a unified approach to working with different data sources.
 There are two sources of Australia-wide climate and weather station observations and data, BoM and Queensland Government's Scientific Information for Landowners (SILO) [@Jeffrey2001] database, hosted by the Queensland Department of the Environment, Science and Innovation (DESI) and one source for Western Australia only weather station data, the Department of Primary Industries and Regional Development of Western Australia's (DPIRD) database.
 BoM provides files through their public anonymous FTP server [@BOM2024] and data from both the SILO and DPIRD databases are available through APIs.
 
-{weatherOz} provides access to data from the SILO database are made available under a Creative Commons Attribution 4.0 International (CC BY 4.0) licence including the Patched Point data available "from 1889 to yesterday" for approximately 8,000 weather station locations located at post-offices, airports, police stations, national parks and private properties [@SILO2024], and Data Drill data, which are spatially interpolated data covering Australian land surfaces.
+`weatherOz` provides access to data from the SILO database are made available under a Creative Commons Attribution 4.0 International (CC BY 4.0) licence including the Patched Point data available "from 1889 to yesterday" for approximately 8,000 weather station locations located at post-offices, airports, police stations, national parks and private properties [@SILO2024], and Data Drill data, which are spatially interpolated data covering Australian land surfaces.
 The DPIRD weather station network includes approximately 200 weather stations with data available from the year 2000 in Western Australia with the majority being located in the south-western part of the state and are available in time-steps from minute to annual summaries with the default being daily values and is made available under a Creative Commons Attribution Licence 3.0 (CC BY 3.0 AU), though users must register for a free API key to use the resource.
 BoM data that are supported include forecasts, both précis and coastal, agriculture bulletins and satellite and radar imagery and are available under a Creative Commons Attribution (CC BY) licence.
 
@@ -96,12 +96,12 @@ Additionally, each data source has a dedicated vignette that details how to fetc
 Most users will likely use `get_patched_point()` (station data), `get_data_drill()` (spatially interpolated gridded data) or `get_dpird_summaries()` to retrieve summarised weather station data values.
[…]
