### 2024-06-03 f563a04
Adds "(BOM)" to metadata for consistency with other station owners
