+++ vignettes/weatherOz.Rmd
@@ -50,17 +50,6 @@ This example assumes that you have stored your DPIRD API key in your .Renviron f
 
 ``` r
 library(weatherOz)
-#> 
-#> Attaching package: 'weatherOz'
-#> The following object is masked from 'package:graphics':
-#> 
-#>     plot
-#> The following object is masked from 'package:base':
-#> 
-#>     plot
-```
-
-``` r
 
 Northam <-
   find_nearby_stations(
@@ -72,10 +61,14 @@ Northam <-
   )
 
 Northam
-#>    station_code station_name longitude  latitude  state elev_m                                                                owner distance_km
-#>          <fctr>       <char>     <num>     <num> <char>  <int>                                                               <char>       <num>
-#> 1:           NO      Northam  116.6942 -31.65161     WA    163 WA Department of Primary Industries and Regional Development (DPIRD)        3.23
-#> 2:           MK       Muresk  116.6913 -31.72772     WA    251 WA Department of Primary Industries and Regional Development (DPIRD)        7.93
+#>    station_code station_name longitude  latitude  state elev_m
+#>          <fctr>       <char>     <num>     <num> <char>  <int>
+#> 1:           NO      Northam  116.6942 -31.65161     WA    163
+#> 2:           MK       Muresk  116.6913 -31.72772     WA    251
+#>                                                                   owner distance_km
+#>                                                                  <char>       <num>
+#> 1: WA Department of Primary Industries and Regional Development (DPIRD)        3.23
+#> 2: WA Department of Primary Industries and Regional Development (DPIRD)        7.93
 ```
 
 There are two stations within 20km of the coordinates for Northam, WA that we provided.
@@ -93,19 +86,45 @@ dpird_daily <- get_dpird_summaries(
 
 dpird_daily
 #> Key: <station_code>
-#>      station_code station_name longitude  latitude  year month   day       date air_tavg air_tmax       air_tmax_time air_tmin       air_tmin_time rainfall rh_avg rh_tmax        rh_tmax_time rh_tmin        rh_tmin_time
-#>            <fctr>       <char>     <num>     <num> <int> <int> <int>     <Date>    <num>    <num>              <POSc>    <num>              <POSc>    <num>  <num>   <num>              <POSc>   <num>              <POSc>
-#>   1:           NO      Northam  116.6942 -31.65161  2022     1     1 2022-01-01     23.8     32.5 2022-01-01 17:02:00     16.7 2022-01-01 05:17:00        0   39.9    71.5 2022-01-01 04:10:00    14.5 2022-01-01 18:03:00
-#>   2:           NO      Northam  116.6942 -31.65161  2022     1     2 2022-01-02     25.4     36.1 2022-01-02 16:58:00     14.4 2022-01-02 05:19:00        0   35.2    63.8 2022-01-02 05:39:00    16.4 2022-01-02 16:58:00
-#>   3:           NO      Northam  116.6942 -31.65161  2022     1     3 2022-01-03     28.0     39.9 2022-01-03 14:42:00     16.6 2022-01-03 05:18:00        0   38.0    68.5 2022-01-03 05:22:00    14.4 2022-01-03 16:15:00
-#>   4:           NO      Northam  116.6942 -31.65161  2022     1     4 2022-01-04     26.1     36.3 2022-01-04 15:15:00     17.3 2022-01-04 04:22:00        0   45.9    74.2 2022-01-04 04:33:00    20.1 2022-01-04 16:35:00
-#>   5:           NO      Northam  116.6942 -31.65161  2022     1     5 2022-01-05     27.3     39.1 2022-01-05 16:19:00     16.1 2022-01-05 05:14:00        0   42.1    77.2 2022-01-05 05:22:00    14.4 2022-01-05 17:21:00
[…]
