### 2024-06-03 5f4a85b
Add check/conversion from fct to char for station_code
