+++ R/get_patched_point.R
@@ -172,6 +172,12 @@ get_patched_point <- function(station_code,
                               values = "all",
                               api_key) {
 
+  # simplify using the metadata to fetch weather data by converting factors to
+  # numeric values
+  if (inherits(x = station_code, what = "factor")) {
+    station_code <- as.character(station_code)
+  }
+
   if (missing(station_code) | !is.character(station_code)) {
     stop(call. = FALSE,
          "Please supply a valid `station_code`.")
