+++ R/get_dpird_extremes.R
@@ -109,6 +109,12 @@ get_dpird_extremes <- function(station_code,
                                 values = "all",
                                 include_closed = FALSE,
                                 api_key) {
+  # simplify using the metadata to fetch weather data by converting factors to
+  # numeric values
+  if (inherits(x = station_code, what = "factor")) {
+    station_code <- as.character(station_code)
+  }
+
   if (missing(station_code) | !is.character(station_code)) {
     stop(
       call. = FALSE,
