+++ R/get_dpird_summaries.R
@@ -228,6 +228,12 @@ get_dpird_summaries <- function(station_code,
     }
   }
 
+  # simplify using the metadata to fetch weather data by converting factors to
+  # numeric values
+  if (inherits(x = station_code, what = "factor")) {
+    station_code <- as.character(station_code)
+  }
+
   if (missing(station_code) | !is.character(station_code)) {
     stop(call. = FALSE,
          "Please supply a valid `station_code`.")
