+++ R/get_patched_point_apsim.R
@@ -113,6 +113,13 @@ get_patched_point_apsim <- function(station_code,
                                     start_date,
                                     end_date = Sys.Date(),
                                     api_key) {
+
+  # simplify using the metadata to fetch weather data by converting factors to
+  # numeric values
+  if (inherits(x = station_code, what = "factor")) {
+    station_code <- as.character(station_code)
+  }
+
   if (missing(station_code) | !is.character(station_code)) {
     stop(call. = FALSE,
          "Please supply a valid `station_code`.")
