### 2024-06-03 6e5c6e3
Redocument to include station_code as factor
