+++ R/get_dpird_minute.R
@@ -5,9 +5,9 @@
 #' Fetch nicely formatted minute weather station data from the \acronym{DPIRD}
 #'   Weather 2.0 \acronym{API} for a maximum 24-hour period.
 #'
-#' @param station_code A `character` string or `vector` of the \acronym{DPIRD}
-#'   station code for the station of interest.  Station codes are available from
-#'   the `get_stations_metadata()` function.
+#' @param station_code A `character` string or `factor` from
+#'   [get_stations_metadata()] of the \acronym{BOM} station code for the station
+#'   of interest.
 #' @param start_date_time A `character` string representing the start date and
 #'   time of the query in the format \dQuote{yyyy-mm-dd-hh-mm} (ISO8601).
 #'   Defaults to 24 hours before the current local system time, returning the
