+++ R/get_dpird_summaries.R
@@ -21,9 +21,9 @@
 #'    \acronym{API} only provide access to daily data, so don't mix (sub)hourly,
 #'    monthly or yearly data from \acronym{DPIRD} with \acronym{SILO}.
 #'
-#' @param station_code A `character` string of the \acronym{DPIRD} station code
-#'   for the station of interest.  Station codes are available from the
-#'   `get_stations_metadata()` function.
+#' @param station_code A `character` string or `factor` from
+#'   [get_stations_metadata()] of the \acronym{BOM} station code for the station
+#'   of interest.
 #' @param start_date A `character` string or `Date` object representing the
 #'   beginning of the range to query in the format \dQuote{yyyy-mm-dd}
 #'   (ISO8601).  Data returned is inclusive of this date.
