+++ man/get_dpird_minute.Rd
@@ -13,9 +13,9 @@ get_dpird_minute(
 )
 }
 \arguments{
-\item{station_code}{A \code{character} string or \code{vector} of the \acronym{DPIRD}
-station code for the station of interest.  Station codes are available from
-the \code{get_stations_metadata()} function.}
+\item{station_code}{A \code{character} string or \code{factor} from
+\code{\link[=get_stations_metadata]{get_stations_metadata()}} of the \acronym{BOM} station code for the station
+of interest.}
 
 \item{start_date_time}{A \code{character} string representing the start date and
 time of the query in the format \dQuote{yyyy-mm-dd-hh-mm} (ISO8601).
