### 2024-06-03 fe83cd0
Check if user possibly supplied closed station without including
