### 2024-06-03 d4192a8
fixes bug where `include_closed` wasn't passed to `get_stations_metadata()`.
