### 2024-06-04 #6717 @bozaah
Thanks @arfon, we have [updated the manuscript](https://github.com/ropensci/weatherOz/commit/137f6feb5acd80af02f40127ef9077a3c5013286) to have the package names use backticks and removed the `{}` (thanks @adamhsparks!). It should be all good now. Thank you.
