### 2024-06-04 e241057
Fix a foot-gun and assume all DPIRD open/closed stations and fix docs
