+++ man/get_dpird_summaries.Rd
@@ -10,7 +10,6 @@ get_dpird_summaries(
   end_date = Sys.Date(),
   interval = c("daily", "15min", "30min", "hourly", "monthly", "yearly"),
   values = "all",
-  include_closed = FALSE,
   api_key
 )
 }
@@ -37,12 +36,6 @@ the previous year.}
 \item{values}{A \code{character} string with the type of summarised weather
 to return.  See \strong{Available Values} for a full list of valid values.
 Defaults to \code{all} with all available values being returned.}
-
-\item{include_closed}{A \code{Boolean} value that defaults to \code{FALSE}.  If set to
-\code{TRUE} the query returns closed and open stations.  Closed stations are
-those that have been turned off and no longer report data.  They may be
-useful for historical purposes.  Only set to \code{TRUE} to fetch data from
-closed stations.}
 
 \item{api_key}{A \code{character} string containing your \acronym{API} key from
 \acronym{DPIRD}, \url{https://www.agric.wa.gov.au/web-apis}, for the
