### 2024-06-04 7bdb2a0
Removes footgun with closed stations. Just ask for data from all.
