+++ R/get_dpird_summaries.R
@@ -40,11 +40,6 @@
 #' @param values A `character` string with the type of summarised weather
 #'   to return.  See **Available Values** for a full list of valid values.
 #'   Defaults to `all` with all available values being returned.
-#' @param include_closed A `Boolean` value that defaults to `FALSE`.  If set to
-#'   `TRUE` the query returns closed and open stations.  Closed stations are
-#'   those that have been turned off and no longer report data.  They may be
-#'   useful for historical purposes.  Only set to `TRUE` to fetch data from
-#'   closed stations.
 #' @param api_key A `character` string containing your \acronym{API} key from
 #'   \acronym{DPIRD}, <https://www.agric.wa.gov.au/web-apis>, for the
 #'   \acronym{DPIRD} Weather 2.0 \acronym{API}.
@@ -195,9 +190,7 @@ get_dpird_summaries <- function(station_code,
                                              "monthly",
                                              "yearly"),
                                 values = "all",
-                                include_closed = FALSE,
                                 api_key) {
-
   # this section is necessary to double check availability dates of DPIRD
   # stations to give users a better experience. It slows down the first query
   # but avoids confusion when no values are returned for a station that exists
@@ -207,23 +200,26 @@ get_dpird_summaries <- function(station_code,
 
     if (!file.exists(metadata_file)) {
       saveRDS(
-        get_stations_metadata(which_api = "dpird",
-                              api_key = api_key,
-                              include_closed = include_closed),
+        get_stations_metadata(
+          which_api = "dpird",
+          api_key = api_key,
+          include_closed = TRUE
+        ),
         file = metadata_file,
         compress = FALSE
       )
     }
 
   } else {
-
     metadata_file <- file.path(tempdir(), "dpird_metadata.Rda")
 
     if (!file.exists(metadata_file)) {
       saveRDS(
-        get_stations_metadata(which_api = "dpird",
-                              api_key = api_key,
-                              include_closed = include_closed),
+        get_stations_metadata(
+          which_api = "dpird",
+          api_key = api_key,
+          include_closed = TRUE
+        ),
         file = metadata_file,
         compress = FALSE
       )
@@ -237,13 +233,11 @@ get_dpird_summaries <- function(station_code,
   }
 
   if (missing(station_code) | !is.character(station_code)) {
-    stop(call. = FALSE,
-         "Please supply a valid `station_code`.")
+    stop(call. = FALSE, "Please supply a valid `station_code`.")
   }
 
   if (missing(start_date))
-    stop(call. = FALSE,
-         "Please supply a valid start date as `start_date`.")
+    stop(call. = FALSE, "Please supply a valid start date as `start_date`.")
 
   if (any(values == "all")) {
     values <-
@@ -253,8 +247,7 @@ get_dpird_summaries <- function(station_code,
         weatherOz::dpird_summary_values)
   } else {
     if (any(values %notin% weatherOz::dpird_summary_values)) {
-      stop(call. = FALSE,
-           "You have specified invalid weather values.")
+      stop(call. = FALSE, "You have specified invalid weather values.")
     }
     values <-
       c("stationCode", "stationName", "period", values)
@@ -264,45 +257,27 @@ get_dpird_summaries <- function(station_code,
   start_date <- .check_date(start_date)
   end_date <- .check_date(end_date)
   .check_date_order(start_date, end_date)
-  tryCatch(
-    expr = {
-      .check_earliest_available_dpird(station_code, start_date, metadata_file)
-    },
-    error = function(e) {
-      message(
-        "Error: have you perhaps supplied a closed station without specifying,",
[…]
