### 2024-06-04 c821448
Reknit vignettes with working file downloads
