+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -168,19 +168,32 @@ library(weatherOz)
   )
 )
 #> Key: <station_code>
-#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall dew_point wet_bulb wind_height wind_avg_speed wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <num>     <num>    <num>       <int>          <num>                           <char>                      <int>          <num>          <num>
-#>    1:           NO 2024-05-25 13:20:00            22.9              58.9             23.0              584        0      14.4     18.1           3          10.73                               NW                        322          6.192          15.73
-#>    2:           NO 2024-05-25 13:21:00            23.1              58.9             23.0              584        0      14.6     18.3           3           9.72                               NW                        318          7.200          14.22
-#>    3:           NO 2024-05-25 13:22:00            23.3              58.0             23.1              585        0      14.6     18.4           3          10.22                               NW                        326          8.712          13.21
-#>    4:           NO 2024-05-25 13:23:00            23.3              57.5             23.1              581        0      14.4     18.3           3           7.20                               NW                        312          4.680          11.23
-#>    5:           NO 2024-05-25 13:24:00            23.2              58.3             23.2              583        0      14.6     18.3           3           9.72                               NW                        310          6.696          13.72
-#>   ---                                                                                                                                                                                                                                                      
-#> 1432:           NO 2024-05-26 13:11:00            24.4              44.6             25.7              646        0      11.6     17.3           3           7.20                               NW                        326          5.688          12.20
-#> 1433:           NO 2024-05-26 13:12:00            24.5              44.5             25.8              640        0      11.6     17.3           3           6.19                               NW                        314          4.680           8.21
-#> 1434:           NO 2024-05-26 13:13:00            24.6              43.4             25.8              634        0      11.3     17.2           3           6.19                               NW                        310          4.176          10.22
-#> 1435:           NO 2024-05-26 13:14:00            24.9              43.1             25.8              628        0      11.5     17.4           3           7.20                               NW                        323          4.176          11.23
-#> 1436:           NO 2024-05-26 13:15:00            25.0              41.7             25.9              621        0      11.1     17.3           3          10.73                               NW                        324          7.200          14.72
+#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall dew_point wet_bulb
+#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <int>     <num>    <num>
+#>    1:           NO 2024-06-03 17:19:00            18.0              56.2             18.0                1        0       9.2     13.6
+#>    2:           NO 2024-06-03 17:20:00            18.0              56.4             18.0                1        0       9.2     13.6
+#>    3:           NO 2024-06-03 17:21:00            18.0              57.8             18.0                1        0       9.6     13.7
+#>    4:           NO 2024-06-03 17:22:00            17.9              56.3             17.9                1        0       9.1     13.5
+#>    5:           NO 2024-06-03 17:23:00            17.9              57.2             17.9                0        0       9.3     13.6
+#>   ---                                                                                                                                 
[…]
