+++ vignettes/weatherOz_for_SILO.Rmd
@@ -105,45 +105,71 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude  year month   day       date air_tmax air_tmax_source air_tmin air_tmin_source  elev_m et_morton_actual et_morton_actual_source et_morton_potential et_morton_potential_source et_morton_wet et_morton_wet_source et_short_crop et_short_crop_source
-#>          <num>    <num> <num> <num> <int>     <Date>    <num>           <int>    <num>           <int>  <char>            <num>                   <int>               <num>                      <int>         <num>                <int>         <num>                <int>
-#>   1:    150.05   -27.85  2022     1     1 2022-01-01     29.7              25     18.8              25 254.5 m              4.3                      26                 7.2                         26           5.7                   26           5.2                   26
-#>   2:    150.05   -27.85  2022     1     2 2022-01-02     33.6              25     16.6              25 254.5 m              6.2                      26                 9.6                         26           7.9                   26           7.0                   26
-#>   3:    150.05   -27.85  2022     1     3 2022-01-03     35.7              25     18.1              25 254.5 m              4.3                      26                11.5                         26           7.9                   26           7.7                   26
-#>   4:    150.05   -27.85  2022     1     4 2022-01-04     36.3              25     18.1              25 254.5 m              5.1                      26                11.1                         26           8.1                   26           7.7                   26
-#>   5:    150.05   -27.85  2022     1     5 2022-01-05     36.4              25     20.3              25 254.5 m              5.0                      26                10.9                         26           8.0                   26           7.5                   26
-#>  ---                                                                                                                                                                                                                                                                        
-#> 872:    150.05   -27.85  2024     5    21 2024-05-21     22.2              25      2.6              25 254.5 m              1.0                      26                 3.6                         26           2.3                   26           2.6                   26
-#> 873:    150.05   -27.85  2024     5    22 2024-05-22     23.3              25      5.1              25 254.5 m              1.2                      26                 3.6                         26           2.4                   26           2.6                   26
-#> 874:    150.05   -27.85  2024     5    23 2024-05-23     22.8              25      4.6              25 254.5 m              1.4                      26                 3.4                         26           2.4                   26           2.5                   26
-#> 875:    150.05   -27.85  2024     5    24 2024-05-24     22.1              25      4.7              25 254.5 m              0.9                      26                 2.9                         26           1.9                   26           2.2                   26
-#> 876:    150.05   -27.85  2024     5    25 2024-05-25     21.8              75      7.2              25 254.5 m              1.1                      26                 2.0                         26           1.6                   26           1.8                   26
-#>      et_tall_crop et_tall_crop_source evap_comb evap_comb_source evap_morton_lake evap_morton_lake_source evap_pan evap_pan_source evap_syn evap_syn_source  extracted   mslp mslp_source radiation radiation_source rainfall rainfall_source rh_tmax rh_tmax_source rh_tmin
-#>             <num>               <int>     <num>            <int>            <num>                   <int>    <num>           <int>    <num>           <int>     <Date>  <num>       <int>     <num>            <int>    <num>           <int>   <num>          <int>   <num>
-#>   1:          6.3                  26       5.0               25              6.0                      26      5.0              25      6.8              26 2024-05-26 1011.8          25      22.6               42     54.9              25    48.0             26    92.2
-#>   2:          8.6                  26       9.2               25              8.3                      26      9.2              25      9.2              26 2024-05-26 1008.5          25      31.3               42      0.1              25    36.3             26   100.0
[…]
