+++ R/parse_ag_bulletin.R
@@ -37,7 +37,7 @@
 #' # parse the ag bulletin for Western Australia
 #'
 #' # download to tempfile() using basename() to keep original name
-#' download.file(url = "ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ60604.xml",
+#' utils::download.file(url = "ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ60604.xml",
 #'               destfile = file.path(tempdir(),
 #'               basename("ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ60604.xml")),
 #'               mode = "wb")
