### 2024-06-04 de27173
Replace curl fns with download.file (#75)
