### 2024-06-05 305bbb0
set timeout value for BOM's FTP server
