+++ R/get_precis_forecast.R
@@ -133,6 +133,9 @@ get_precis_forecast <- function(state = "AUS") {
 
 .parse_precis_forecast <- function(xml_url) {
 
+  op <- options(timeout = 120L)
+  on.exit(options(op))
+
   # load the XML from ftp
   if (substr(xml_url, 1, 3) == "ftp") {
     xml_object <- .get_url(xml_url)
