### 2024-06-05 0cb95c9
Use an unprojected `sf` object to avoid UTF8 NOTES in checks
