+++ R/data.R
@@ -87,7 +87,8 @@
 #' 1:20,000
 #'
 #' @section Coordinate Reference System:
-#' EPSG:28350 - GDA94 / MGA zone 50 \url{https://epsg.io/28350}
+#' EPSG:4326 - WGS 84 -- WGS84 - World Geodetic System 1984, used in GPS
+#'   <https://epsg.io/4326>
 #'
 #' @name south_west_agriculture_region
 #'
