### 2024-06-05 7937f27
Update NEWS.md
