### 2024-06-05 de601ec
Fix links in README for CRAN
