+++ README.md
@@ -13,7 +13,7 @@ Review](https://badges.ropensci.org/598_status.svg)](https://github.com/ropensci
 [![status](https://joss.theoj.org/papers/44a93e6baee10f5212735041123d400b/status.svg)](https://joss.theoj.org/papers/44a93e6baee10f5212735041123d400b)
 [![R-CMD-check](https://github.com/ropensci/weatherOz/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ropensci/weatherOz/actions/workflows/R-CMD-check.yaml)
 [![test-coverage](https://github.com/ropensci/weatherOz/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/ropensci/weatherOz/actions/workflows/test-coverage.yaml)
-[![codecov](https://codecov.io/gh/ropensci/weatherOz/graph/badge.svg?token=ZiaPsN6nYy)](https://codecov.io/gh/ropensci/weatherOz)
+[![codecov](https://codecov.io/gh/ropensci/weatherOz/graph/badge.svg?token=ZiaPsN6nYy)](https://app.codecov.io/gh/ropensci/weatherOz)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/weatherOz)](https://CRAN.R-project.org/package=weatherOz)
 
@@ -95,6 +95,9 @@ library(weatherOz)
 #> The following object is masked from 'package:base':
 #> 
 #>     plot
+```
+
+``` r
 
 wd <- get_dpird_summaries(
   station_code = "BI",
@@ -170,12 +173,12 @@ head(wd)
 #> 6:    150.05   -27.85  2022    10     6 2022-10-06     24.4              25
 #>    air_tmin air_tmin_source  elev_m  extracted rainfall rainfall_source
 #>       <num>           <int>  <char>     <Date>    <num>           <int>
-#> 1:      9.8              25 254.5 m 2024-05-04      0.9              25
-#> 2:     11.7              25 254.5 m 2024-05-04      0.0              25
-#> 3:      7.8              25 254.5 m 2024-05-04      0.0              25
-#> 4:     10.6              25 254.5 m 2024-05-04      0.0              25
-#> 5:     13.3              25 254.5 m 2024-05-04      0.0              25
-#> 6:     14.7              25 254.5 m 2024-05-04      1.8              25
+#> 1:      9.8              25 254.5 m 2024-06-05      0.9              25
+#> 2:     11.7              25 254.5 m 2024-06-05      0.0              25
+#> 3:      7.8              25 254.5 m 2024-06-05      0.0              25
+#> 4:     10.6              25 254.5 m 2024-06-05      0.0              25
+#> 5:     13.3              25 254.5 m 2024-06-05      0.0              25
+#> 6:     14.7              25 254.5 m 2024-06-05      1.8              25
 ```
 
 ## Notes on Data and API Endpoints
