### 2024-06-05 e64bd58
Update CRAN comments for new release
