+++ cran-comments.md
@@ -2,4 +2,10 @@
 
 0 errors | 0 warnings | 1 note
 
+## URL Check Note
+
+Several URLs are reported as "Forbidden".
+This is due to an Australian Bureau of Meteorology policy that blocks any headless access of web resources that are served via HTTP methods but if accessed via a browser, properly respond.
+The URLs have all been validated and are considered to be stable and unchanging as a part of a government website.
+
 * This is a new release.
