### 2024-06-05 1868678
Reknit vignettes to fix issue with sw-ag-region map
