+++ vignettes/weatherOz_for_SILO.Rmd
@@ -113,11 +113,11 @@ library(weatherOz)
 #>   4:    150.05   -27.85  2022     1     4 2022-01-04     36.3              25     18.1              25 254.5 m              5.1
 #>   5:    150.05   -27.85  2022     1     5 2022-01-05     36.4              25     20.3              25 254.5 m              5.0
 #>  ---                                                                                                                           
-#> 881:    150.05   -27.85  2024     5    30 2024-05-30     24.9              25      9.9              25 254.5 m              1.4
 #> 882:    150.05   -27.85  2024     5    31 2024-05-31     23.7              25     14.1              25 254.5 m              1.0
 #> 883:    150.05   -27.85  2024     6     1 2024-06-01     18.4              25     16.4              25 254.5 m              1.4
 #> 884:    150.05   -27.85  2024     6     2 2024-06-02     17.8              25      8.2              25 254.5 m              1.5
-#> 885:    150.05   -27.85  2024     6     3 2024-06-03     20.7              75      5.1              25 254.5 m              1.3
+#> 885:    150.05   -27.85  2024     6     3 2024-06-03     17.0              25      5.1              25 254.5 m              1.6
+#> 886:    150.05   -27.85  2024     6     4 2024-06-04     20.6              75      2.0              25 254.5 m              1.2
 #>      et_morton_actual_source et_morton_potential et_morton_potential_source et_morton_wet et_morton_wet_source et_short_crop
 #>                        <int>               <num>                      <int>         <num>                <int>         <num>
 #>   1:                      26                 7.2                         26           5.7                   26           5.2
@@ -126,50 +126,50 @@ library(weatherOz)
 #>   4:                      26                11.1                         26           8.1                   26           7.7
 #>   5:                      26                10.9                         26           8.0                   26           7.5
 #>  ---                                                                                                                        
-#> 881:                      26                 3.2                         26           2.3                   26           2.3
 #> 882:                      26                 3.0                         26           2.0                   26           2.1
 #> 883:                      26                 1.4                         26           1.4                   26           1.1
 #> 884:                      26                 2.3                         26           1.9                   26           1.6
-#> 885:                      26                 2.9                         26           2.1                   26           2.1
-#>      et_short_crop_source et_tall_crop et_tall_crop_source evap_comb evap_comb_source evap_morton_lake evap_morton_lake_source evap_pan
-#>                     <int>        <num>               <int>     <num>            <int>            <num>                   <int>    <num>
-#>   1:                   26          6.3                  26       5.0               25              6.0                      26      5.0
-#>   2:                   26          8.6                  26       9.2               25              8.3                      26      9.2
-#>   3:                   26          9.9                  26       8.1               25              8.3                      26      8.1
-#>   4:                   26          9.7                  26       8.6               25              8.5                      26      8.6
-#>   5:                   26          9.4                  26       6.3               25              8.3                      26      6.3
-#>  ---                                                                                                                                   
-#> 881:                   26          3.2                  26       2.2               25              2.3                      26      2.2
-#> 882:                   26          2.8                  26       1.7               25              2.0                      26      1.7
[…]
