+++ vignettes/weatherOz_for_BOM.Rmd
@@ -66,43 +66,43 @@ library(weatherOz)
 #>   --------------------------------------------------------------------------------  
 #>       index product_id  state         town       aac       lat      lon  elev    start_time_local end_time_local utc_offset
 #>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>              <POSc>         <POSc>     <fctr>
-#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-04 17:00:00     2024-06-05      10:00
-#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-05 00:00:00     2024-06-06      10:00
-#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-06 00:00:00     2024-06-07      10:00
-#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-07 00:00:00     2024-06-08      10:00
-#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-08 00:00:00     2024-06-09      10:00
+#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-05 17:00:00     2024-06-06      10:00
+#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-06 00:00:00     2024-06-07      10:00
+#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-07 00:00:00     2024-06-08      10:00
+#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-08 00:00:00     2024-06-09      10:00
+#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-09 00:00:00     2024-06-10      10:00
 #>  ---                                                                                                                       
-#> 900:      3   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-07 00:00:00     2024-06-08      10:00
-#> 901:      4   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-08 00:00:00     2024-06-09      10:00
-#> 902:      5   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-09 00:00:00     2024-06-10      10:00
-#> 903:      6   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-10 00:00:00     2024-06-11      10:00
-#> 904:      7   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-11 00:00:00     2024-06-12      10:00
-#>           start_time_utc        end_time_utc minimum_temperature maximum_temperature lower_precipitation_limit upper_precipitation_limit
-#>                   <POSc>              <POSc>               <num>               <num>                     <num>                     <num>
-#>   1: 2024-06-04 07:00:00 2024-06-04 14:00:00                  NA                  NA                        NA                        NA
-#>   2: 2024-06-04 14:00:00 2024-06-05 14:00:00                   8                  22                        NA                        NA
-#>   3: 2024-06-05 14:00:00 2024-06-06 14:00:00                  12                  22                         0                         1
[…]

[…]
