### 2024-06-05 6db2c71
Ensure images appear in DPIRD vignette
