+++ vignettes/weatherOz.Rmd
@@ -50,6 +50,17 @@ This example assumes that you have stored your DPIRD API key in your .Renviron f
 
 ``` r
 library(weatherOz)
+#> 
+#> Attaching package: 'weatherOz'
+#> The following object is masked from 'package:graphics':
+#> 
+#>     plot
+#> The following object is masked from 'package:base':
+#> 
+#>     plot
+```
+
+``` r
 
 Northam <-
   find_nearby_stations(
@@ -86,32 +97,45 @@ dpird_daily <- get_dpird_summaries(
 
 dpird_daily
 #> Key: <station_code>
-#>      station_code station_name longitude  latitude  year month   day       date air_tavg air_tmax       air_tmax_time air_tmin
-#>            <fctr>       <char>     <num>     <num> <int> <int> <int>     <Date>    <num>    <num>              <POSc>    <num>
-#>   1:           NO      Northam  116.6942 -31.65161  2022     1     1 2022-01-01     23.8     32.5 2022-01-01 17:02:00     16.7
-#>   2:           NO      Northam  116.6942 -31.65161  2022     1     2 2022-01-02     25.4     36.1 2022-01-02 16:58:00     14.4
-#>   3:           NO      Northam  116.6942 -31.65161  2022     1     3 2022-01-03     28.0     39.9 2022-01-03 14:42:00     16.6
-#>   4:           NO      Northam  116.6942 -31.65161  2022     1     4 2022-01-04     26.1     36.3 2022-01-04 15:15:00     17.3
-#>   5:           NO      Northam  116.6942 -31.65161  2022     1     5 2022-01-05     27.3     39.1 2022-01-05 16:19:00     16.1
-#>  ---                                                                                                                          
-#> 361:           NO      Northam  116.6942 -31.65161  2022    12    27 2022-12-27     23.0     40.2 2022-12-27 14:48:20     13.7
-#> 362:           NO      Northam  116.6942 -31.65161  2022    12    28 2022-12-28     22.2     40.5 2022-12-28 19:58:20     11.9
-#> 363:           NO      Northam  116.6942 -31.65161  2022    12    29 2022-12-29     26.0     40.2 2022-12-30 03:03:20     13.8
-#> 364:           NO      Northam  116.6942 -31.65161  2022    12    30 2022-12-30     26.6     39.2 2022-12-30 13:50:50     14.5
-#> 365:           NO      Northam  116.6942 -31.65161  2022    12    31 2022-12-31     22.7     31.1 2022-12-31 15:45:50     16.2
-#>            air_tmin_time rainfall rh_avg rh_tmax        rh_tmax_time rh_tmin        rh_tmin_time
-#>                   <POSc>    <num>  <num>   <num>              <POSc>   <num>              <POSc>
-#>   1: 2022-01-01 05:17:00        0   39.9    71.5 2022-01-01 04:10:00    14.5 2022-01-01 18:03:00
-#>   2: 2022-01-02 05:19:00        0   35.2    63.8 2022-01-02 05:39:00    16.4 2022-01-02 16:58:00
-#>   3: 2022-01-03 05:18:00        0   38.0    68.5 2022-01-03 05:22:00    14.4 2022-01-03 16:15:00
-#>   4: 2022-01-04 04:22:00        0   45.9    74.2 2022-01-04 04:33:00    20.1 2022-01-04 16:35:00
-#>   5: 2022-01-05 05:14:00        0   42.1    77.2 2022-01-05 05:22:00    14.4 2022-01-05 17:21:00
-#>  ---                                                                                            
[…]
