+++ vignettes/weatherOz_for_BOM.Rmd
@@ -64,45 +64,58 @@ library(weatherOz)
 #>   Please note information at the page 
 #>   <http://www.bom.gov.au/catalogue/data-feeds.shtml#precis>.
 #>   --------------------------------------------------------------------------------  
-#>       index product_id  state         town       aac       lat      lon  elev    start_time_local end_time_local utc_offset
-#>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>              <POSc>         <POSc>     <fctr>
-#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-05 17:00:00     2024-06-06      10:00
-#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-06 00:00:00     2024-06-07      10:00
-#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-07 00:00:00     2024-06-08      10:00
-#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-08 00:00:00     2024-06-09      10:00
-#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-06-09 00:00:00     2024-06-10      10:00
-#>  ---                                                                                                                       
-#> 900:      3   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-08 00:00:00     2024-06-09      10:00
-#> 901:      4   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-09 00:00:00     2024-06-10      10:00
-#> 902:      5   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-10 00:00:00     2024-06-11      10:00
-#> 903:      6   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-11 00:00:00     2024-06-12      10:00
-#> 904:      7   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-06-12 00:00:00     2024-06-13      10:00
-#>           start_time_utc        end_time_utc minimum_temperature maximum_temperature lower_precipitation_limit
-#>                   <POSc>              <POSc>               <num>               <num>                     <num>
-#>   1: 2024-06-05 07:00:00 2024-06-05 14:00:00                  NA                  NA                        NA
-#>   2: 2024-06-05 14:00:00 2024-06-06 14:00:00                  11                  23                        NA
-#>   3: 2024-06-06 14:00:00 2024-06-07 14:00:00                  10                  21                        NA
-#>   4: 2024-06-07 14:00:00 2024-06-08 14:00:00                   9                  21                        NA
-#>   5: 2024-06-08 14:00:00 2024-06-09 14:00:00                  10                  23                        NA
-#>  ---                                                                                                          
-#> 900: 2024-06-07 14:00:00 2024-06-08 14:00:00                  16                  26                        NA
-#> 901: 2024-06-08 14:00:00 2024-06-09 14:00:00                  15                  26                        NA
-#> 902: 2024-06-09 14:00:00 2024-06-10 14:00:00                  16                  26                        NA
-#> 903: 2024-06-10 14:00:00 2024-06-11 14:00:00                  16                  26                        NA
[…]

[…]
