### 2024-06-05 679df7a
reknit vignettes to fix non-portable names
