+++ vignettes/weatherOz.Rmd
@@ -196,17 +196,17 @@ ppd
 #> 365:       041103    Toowoomba  2022    12    31 2022-12-31     27.7              25     16.4
 #>      air_tmin_source  elev_m  extracted latitude longitude rainfall rainfall_source rh_tmax
 #>                <int>  <char>     <Date>    <num>     <num>    <num>           <int>   <num>
-#>   1:              25 691.0 m 2024-06-05 -27.5836  151.9317      7.6              25    72.1
-#>   2:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.1              25    47.1
-#>   3:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.0              25    38.8
-#>   4:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.0              25    43.1
-#>   5:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.0              25    49.2
+#>   1:              25 691.0 m 2024-06-06 -27.5836  151.9317      7.6              25    72.1
+#>   2:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.1              25    47.1
+#>   3:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.0              25    38.8
+#>   4:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.0              25    43.1
+#>   5:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.0              25    49.2
 #>  ---                                                                                       
-#> 361:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.0              25    47.9
-#> 362:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.2              25    43.7
-#> 363:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.0              25    44.1
-#> 364:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.7              25    60.3
-#> 365:              25 691.0 m 2024-06-05 -27.5836  151.9317      0.4              25    50.9
+#> 361:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.0              25    47.9
+#> 362:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.2              25    43.7
+#> 363:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.0              25    44.1
+#> 364:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.7              25    60.3
+#> 365:              25 691.0 m 2024-06-06 -27.5836  151.9317      0.4              25    50.9
 #>      rh_tmax_source rh_tmin rh_tmin_source
 #>               <int>   <num>          <int>
 #>   1:             26   100.0             26
