+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -170,43 +170,43 @@ library(weatherOz)
 #> Key: <station_code>
 #>       station_code           date_time air_temperature relative_humidity soil_temperature
 #>             <fctr>              <POSc>           <num>             <num>           <lgcl>
-#>    1:           NO 2024-06-04 17:57:00            18.3              72.5               NA
-#>    2:           NO 2024-06-04 17:58:00            18.3              72.2               NA
-#>    3:           NO 2024-06-04 17:59:00            18.2              72.1               NA
-#>    4:           NO 2024-06-04 18:00:00            18.2              71.9               NA
-#>    5:           NO 2024-06-04 18:01:00            18.2              71.7               NA
+#>    1:           NO 2024-06-05 07:15:00            15.0              75.1               NA
+#>    2:           NO 2024-06-05 07:16:00            15.0              75.0               NA
+#>    3:           NO 2024-06-05 07:17:00            15.1              74.7               NA
+#>    4:           NO 2024-06-05 07:18:00            15.1              74.5               NA
+#>    5:           NO 2024-06-05 07:19:00            15.1              74.3               NA
 #>   ---                                                                                    
-#> 1430:           NO 2024-06-05 17:46:00            18.7              52.2               NA
-#> 1431:           NO 2024-06-05 17:47:00            18.7              51.7               NA
-#> 1432:           NO 2024-06-05 17:48:00            18.7              52.0               NA
-#> 1433:           NO 2024-06-05 17:49:00            18.6              52.2               NA
-#> 1434:           NO 2024-06-05 17:50:00            18.6              52.7               NA
+#> 1432:           NO 2024-06-06 07:06:00             9.0              99.9               NA
+#> 1433:           NO 2024-06-06 07:07:00             9.1             100.0               NA
+#> 1434:           NO 2024-06-06 07:08:00             9.1             100.0               NA
+#> 1435:           NO 2024-06-06 07:09:00             9.1             100.0               NA
+#> 1436:           NO 2024-06-06 07:10:00             9.2              99.9               NA
 #>       solar_irradiance rainfall dew_point wet_bulb wind_height wind_avg_speed
 #>                  <int>    <int>     <num>    <num>       <int>          <num>
-#>    1:                0        0      13.3     15.6           3           0.14
-#>    2:                0        0      13.2     15.6           3           0.14
-#>    3:                0        0      13.1     15.5           3           0.14
-#>    4:                0        0      13.1     15.5           3           0.14
-#>    5:                0        0      13.0     15.5           3           0.14
+#>    1:                6        0      10.6     12.8           3           1.66
+#>    2:                8        0      10.6     12.8           3           1.15
+#>    3:                8        0      10.6     12.9           3           0.14
+#>    4:                9        0      10.6     12.9           3           0.14
+#>    5:                9        0      10.6     12.9           3           0.14
 #>   ---                                                                        
-#> 1430:                0        0       8.7     13.7           3           0.14
-#> 1431:                0        0       8.6     13.6           3           0.14
-#> 1432:                0        0       8.7     13.6           3           0.14
-#> 1433:                0        0       8.6     13.6           3           0.14
-#> 1434:                0        0       8.8     13.6           3           0.14
+#> 1432:                3        0       9.0      9.0           3           0.14
+#> 1433:                3        0       9.1      9.1           3           0.14
+#> 1434:                4        0       9.1      9.1           3           0.14
+#> 1435:                4        0       9.1      9.1           3           0.14
+#> 1436:                5        0       9.2      9.2           3           0.14
 #>       wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
 #>                                 <char>                      <int>          <num>          <num>
[…]
