+++ vignettes/weatherOz_for_SILO.Rmd
@@ -113,11 +113,11 @@ library(weatherOz)
 #>   4:    150.05   -27.85  2022     1     4 2022-01-04     36.3              25     18.1              25
 #>   5:    150.05   -27.85  2022     1     5 2022-01-05     36.4              25     20.3              25
 #>  ---                                                                                                  
-#> 882:    150.05   -27.85  2024     5    31 2024-05-31     23.7              25     14.1              25
-#> 883:    150.05   -27.85  2024     6     1 2024-06-01     18.4              25     16.4              25
-#> 884:    150.05   -27.85  2024     6     2 2024-06-02     17.8              25      8.2              25
-#> 885:    150.05   -27.85  2024     6     3 2024-06-03     17.0              25      5.1              25
-#> 886:    150.05   -27.85  2024     6     4 2024-06-04     20.6              75      2.0              25
+#> 883:    150.05   -27.85  2024     6     1 2024-06-01     18.3              25     16.4              25
+#> 884:    150.05   -27.85  2024     6     2 2024-06-02     17.9              25      8.2              25
+#> 885:    150.05   -27.85  2024     6     3 2024-06-03     16.9              25      5.1              25
+#> 886:    150.05   -27.85  2024     6     4 2024-06-04     18.9              25      2.0              25
+#> 887:    150.05   -27.85  2024     6     5 2024-06-05     20.6              75      5.5              25
 #>       elev_m et_morton_actual et_morton_actual_source et_morton_potential et_morton_potential_source
 #>       <char>            <num>                   <int>               <num>                      <int>
 #>   1: 254.5 m              4.3                      26                 7.2                         26
@@ -126,11 +126,11 @@ library(weatherOz)
 #>   4: 254.5 m              5.1                      26                11.1                         26
 #>   5: 254.5 m              5.0                      26                10.9                         26
 #>  ---                                                                                                
-#> 882: 254.5 m              1.0                      26                 3.0                         26
 #> 883: 254.5 m              1.4                      26                 1.4                         26
 #> 884: 254.5 m              1.5                      26                 2.3                         26
 #> 885: 254.5 m              1.6                      26                 2.0                         26
-#> 886: 254.5 m              1.2                      26                 2.7                         26
+#> 886: 254.5 m              1.6                      26                 2.4                         26
+#> 887: 254.5 m              0.4                      26                 1.5                         26
 #>      et_morton_wet et_morton_wet_source et_short_crop et_short_crop_source et_tall_crop
 #>              <num>                <int>         <num>                <int>        <num>
 #>   1:           5.7                   26           5.2                   26          6.3
@@ -139,11 +139,11 @@ library(weatherOz)
 #>   4:           8.1                   26           7.7                   26          9.7
 #>   5:           8.0                   26           7.5                   26          9.4
 #>  ---                                                                                   
-#> 882:           2.0                   26           2.1                   26          2.8
 #> 883:           1.4                   26           1.1                   26          1.3
-#> 884:           1.9                   26           1.6                   26          2.1
+#> 884:           1.9                   26           1.7                   26          2.1
 #> 885:           1.8                   26           1.5                   26          2.0
[…]

[…]
