### 2024-06-06 0acf25a
quote nonstandard words for CRAN check
