### 2024-06-06 7505cfc
Fixes bad URL in SILO vignette
