### 2024-06-06 cbf1900
Reknit vignettes
