+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -168,45 +168,58 @@ library(weatherOz)
   )
 )
 #> Key: <station_code>
-#>       station_code           date_time air_temperature relative_humidity soil_temperature
-#>             <fctr>              <POSc>           <num>             <num>           <lgcl>
-#>    1:           NO 2024-06-05 07:15:00            15.0              75.1               NA
-#>    2:           NO 2024-06-05 07:16:00            15.0              75.0               NA
-#>    3:           NO 2024-06-05 07:17:00            15.1              74.7               NA
-#>    4:           NO 2024-06-05 07:18:00            15.1              74.5               NA
-#>    5:           NO 2024-06-05 07:19:00            15.1              74.3               NA
-#>   ---                                                                                    
-#> 1432:           NO 2024-06-06 07:06:00             9.0              99.9               NA
-#> 1433:           NO 2024-06-06 07:07:00             9.1             100.0               NA
-#> 1434:           NO 2024-06-06 07:08:00             9.1             100.0               NA
-#> 1435:           NO 2024-06-06 07:09:00             9.1             100.0               NA
-#> 1436:           NO 2024-06-06 07:10:00             9.2              99.9               NA
-#>       solar_irradiance rainfall dew_point wet_bulb wind_height wind_avg_speed
-#>                  <int>    <int>     <num>    <num>       <int>          <num>
-#>    1:                6        0      10.6     12.8           3           1.66
-#>    2:                8        0      10.6     12.8           3           1.15
-#>    3:                8        0      10.6     12.9           3           0.14
-#>    4:                9        0      10.6     12.9           3           0.14
-#>    5:                9        0      10.6     12.9           3           0.14
-#>   ---                                                                        
-#> 1432:                3        0       9.0      9.0           3           0.14
-#> 1433:                3        0       9.1      9.1           3           0.14
-#> 1434:                4        0       9.1      9.1           3           0.14
-#> 1435:                4        0       9.1      9.1           3           0.14
-#> 1436:                5        0       9.2      9.2           3           0.14
-#>       wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>                                 <char>                      <int>          <num>          <num>
-#>    1:                              NNE                         19          0.144           3.67
-#>    2:                                N                        358          0.648           2.66
-#>    3:                                N                        358          0.144           1.15
-#>    4:                                N                        358          0.144           0.65
-#>    5:                                N                        358          0.144           1.15
-#>   ---                                                                                          
-#> 1432:                              WSW                        258          0.144           1.15
-#> 1433:                                W                        268          0.144           1.66
-#> 1434:                                W                        269          0.144           0.65
-#> 1435:                                W                        274          0.144           1.15
-#> 1436:                                W                        278          0.144           0.65
+#>       station_code           date_time air_temperature relative_humidity
+#>             <fctr>              <POSc>           <num>             <num>
+#>    1:           NO 2024-06-05 14:35:00            22.5              41.6
+#>    2:           NO 2024-06-05 14:36:00            22.5              41.0
+#>    3:           NO 2024-06-05 14:37:00            22.4              40.9
[…]
