+++ vignettes/weatherOz.Rmd
@@ -76,10 +76,14 @@ Northam
 #>          <fctr>       <char>     <num>     <num> <char>  <int>
 #> 1:           NO      Northam  116.6942 -31.65161     WA    163
 #> 2:           MK       Muresk  116.6913 -31.72772     WA    251
-#>                                                                   owner distance_km
-#>                                                                  <char>       <num>
-#> 1: WA Department of Primary Industries and Regional Development (DPIRD)        3.23
-#> 2: WA Department of Primary Industries and Regional Development (DPIRD)        7.93
+#>                                                                   owner
+#>                                                                  <char>
+#> 1: WA Department of Primary Industries and Regional Development (DPIRD)
+#> 2: WA Department of Primary Industries and Regional Development (DPIRD)
+#>    distance_km
+#>          <num>
+#> 1:        3.23
+#> 2:        7.93
 ```
 
 There are two stations within 20km of the coordinates for Northam, WA that we provided.
@@ -97,32 +101,45 @@ dpird_daily <- get_dpird_summaries(
 
 dpird_daily
 #> Key: <station_code>
-#>      station_code station_name longitude  latitude  year month   day       date air_tavg air_tmax
-#>            <fctr>       <char>     <num>     <num> <int> <int> <int>     <Date>    <num>    <num>
-#>   1:           NO      Northam  116.6942 -31.65161  2022     1     1 2022-01-01     23.8     32.5
-#>   2:           NO      Northam  116.6942 -31.65161  2022     1     2 2022-01-02     25.4     36.1
-#>   3:           NO      Northam  116.6942 -31.65161  2022     1     3 2022-01-03     28.0     39.9
-#>   4:           NO      Northam  116.6942 -31.65161  2022     1     4 2022-01-04     26.1     36.3
-#>   5:           NO      Northam  116.6942 -31.65161  2022     1     5 2022-01-05     27.3     39.1
-#>  ---                                                                                             
-#> 361:           NO      Northam  116.6942 -31.65161  2022    12    27 2022-12-27     23.0     40.2
-#> 362:           NO      Northam  116.6942 -31.65161  2022    12    28 2022-12-28     22.2     40.5
-#> 363:           NO      Northam  116.6942 -31.65161  2022    12    29 2022-12-29     26.0     40.2
-#> 364:           NO      Northam  116.6942 -31.65161  2022    12    30 2022-12-30     26.6     39.2
-#> 365:           NO      Northam  116.6942 -31.65161  2022    12    31 2022-12-31     22.7     31.1
-#>            air_tmax_time air_tmin       air_tmin_time rainfall rh_avg rh_tmax        rh_tmax_time
-#>                   <POSc>    <num>              <POSc>    <num>  <num>   <num>              <POSc>
-#>   1: 2022-01-01 17:02:00     16.7 2022-01-01 05:17:00        0   39.9    71.5 2022-01-01 04:10:00
-#>   2: 2022-01-02 16:58:00     14.4 2022-01-02 05:19:00        0   35.2    63.8 2022-01-02 05:39:00
-#>   3: 2022-01-03 14:42:00     16.6 2022-01-03 05:18:00        0   38.0    68.5 2022-01-03 05:22:00
-#>   4: 2022-01-04 15:15:00     17.3 2022-01-04 04:22:00        0   45.9    74.2 2022-01-04 04:33:00
-#>   5: 2022-01-05 16:19:00     16.1 2022-01-05 05:14:00        0   42.1    77.2 2022-01-05 05:22:00
-#>  ---                                                                                             
[…]

[…]
