### 2024-06-06 #6717 @arfon
@bozaah – can you merge this PR please? It should fix the XML validation above: https://github.com/ropensci/weatherOz/pull/76
