### 2024-06-06 1a84923
Update paper.bib (#76)
