+++ inst/paper/paper.bib
@@ -113,7 +113,7 @@
   bdsk-url-1 = {https://doi.org/10.1094/PHYTO.2003.93.4.428},
   bdsk-url-2 = {http://dx.doi.org/10.1094/PHYTO.2003.93.4.428},
   doi        = {10.1094/PHYTO.2003.93.4.428},
-  isbn       = {0031-949X},
+  issn       = {0031-949X},
   journal1   = {Phytopathology},
   m3         = {doi: 10.1094/PHYTO.2003.93.4.428},
   publisher  = {Scientific Societies},
