### 2024-06-07 6c082dc
Add Paper for JOSS (#77)
Co-authored-by: Dean Marchiori <deanmarchiori@users.noreply.github.com>
Co-authored-by: mark padgham <mark.padgham@email.com>
Co-authored-by: Rodrigo Pires <11730095+bozaah@users.noreply.github.com>
Co-authored-by: Rodrigo Pires <rodrigo.pires@dpird.wa.gov.au>
Co-authored-by: Arfon Smith <arfon@users.noreply.github.com>
