+++ inst/paper/paper.md
@@ -1,0 +1,111 @@
+---
+title:
+  'weatherOz: An API Client for Australian Weather and Climate Data Resources in R'
+authors:
+- affiliation: 1
+  name: Rodrigo Pires
+  orcid: 0000-0001-7384-6849
+- affiliation: 1
+  name: Anna Hepworth
+  orcid: 0000-0003-0204-6347
+- affiliation: 1
+  name: Rebecca O'Leary
+- affiliation: 2
+  name: Jonathan Carroll
+  orcid: 0000-0002-1404-5264
+- affiliation: 3
+  name: James Goldie
+  orcid: 0000-0002-5024-6207
+- affiliation: 4
+  name: Dean Marchiori
+  orcid: 0000-0002-3430-7225
+- affiliation: 5
+  name: Paul Melloy
+  orcid: 0000-0003-4253-7167
+- affiliation: 6
+  name: Mark Padgam
+  orcid: 0000-0003-2172-5265
+- affiliation: 7
+  name: Hugh Parsonage
+  orcid: 0000-0003-4055-0835
+- affiliation: "1, 8, 9"
+  name: Adam H. Sparks
+  orcid: 0000-0002-0061-8359
+date: "25 Mar 2024"
+output:
+  pdf_document: default
+  html_document: default
+bibliography: paper.bib
+tags:
+- Australia
+- weather forecast
+- meteorology
+- climatology
+- weather data
+- R
+- API client
+- xml
+- json
+affiliations:
+- index: 1
+  name: Department of Primary Industries and Regional Development of Western Australia, Perth, Western Australia 6000, Australia
+- index: 2
+  name: No affiliation
+- index: 3
+  name: 360info, Monash University, Australia
+- index: 4
+  name: Wave Data Labs, Wollongong, New South Wales, 2500, Australia
+- index: 5
+  name: The University of Queensland, School of Agriculture and Food Sustainability, Gatton, Queensland 4343, Australia
+- index: 6
+  name: rOpenSci
+- index: 7
+  name: Grattan Institute, Carlton, Victoria 3053, Australia
+- index: 8
+  name: University of Southern Queensland, Centre for Crop Health, Toowoomba, Queensland 4350, Australia
+- index: 9
+  name: Curtin University, Centre for Crop and Disease Management, Curtin Biometry and Agricultural Data Analytics, Bentley, Western Australia 6102, Australia
+---
+
+# Summary
+
+Researchers and policy makers use weather data in a variety of ways.
+Agriculture applications of the data are used in several types of models and decision support tools, to estimate leaf wetness, crop yield, crop growth stage, physiological stress or forecasting crop disease epidemics or insect pest population levels [@Venaelaeinen2002; @DeWolf2003; @Sparks2017].
+Other areas of use include mapping potential renewable energy, *e.g.*, wind or solar potential for exploration purposes [@Ramachandra2007].
+The data can also be used by decision makers for municipalities to help plan for extreme weather events, energy needs and other infrastructure [@Svensson2002; @Alcoforado2009].
+
+`weatherOz` offers R users with a single interface to access Australian climate and weather data sources providing a standardised way of easily querying and retrieving Australian climate and weather data.
+
+# Statement of need
+
+Australian weather data availability is fragmented and difficult to easily access repeatedly or programmatically for use in research activities.
+Previously the R package, `bomrang` [@Sparks2017a], provided easy access to data available from the Australian Bureau of Meteorology (BoM) in R.
+However, difficulties with accessing the data and restrictive polices caused us to archive this package.
+Therefore, we created `weatherOz` to replace and at the same time, improve upon `bomrang` and provide R users with a package that allowed for easy of access to multiple sources of Australian climate and weather data while providing a unified approach to working with different data sources.
[…]
