+++ inst/paper/paper.bib
@@ -1,0 +1,177 @@
+@Webpage{BOM2024,
+  author      = {{Australian Bureau of Meteorology}},
+  lastchecked = {23/02/2024},
+  month       = feb,
+  title       = {Weather Data Services},
+  url         = {http://www.bom.gov.au/catalogue/data-feeds.shtml},
+  year        = {2024},
+}
+
+@Article{Sparks2017a,
+  author    = {Sparks, Adam H. and Padgham, Mark and Parsonage, Hugh and Pembleton, Keith},
+  journal   = {The Journal of Open Source Software},
+  title     = {{bomrang}: {Fetch Australian} Government {Bureau of Meteorology} Data in {R}},
+  year      = {2017},
+  issn      = {2475-9066},
+  month     = sep,
+  number    = {17},
+  pages     = {411},
+  volume    = {2},
+  doi       = {10.21105/joss.00411},
+  publisher = {The Open Journal},
+}
+
+@Manual{Miguez2024,
+  title  = {{apsimx}: Inspect, Read, Edit and Run {'APSIM' "Next Generation"} and {'APSIM' Classic}},
+  author = {Miguez, Fernando},
+  note   = {R package version 2.6.2},
+  year   = {2024},
+  url    = {https://CRAN.R-project.org/package=apsimx},
+}
+
+@Article{Jeffrey2001,
+  author    = {Jeffrey, Stephen J. and Carter, John O. and Moodie, Keith B. and Beswick, Alan R.},
+  journal   = {Environmental Modelling \& Software},
+  title     = {Using spatial interpolation to construct a comprehensive archive of {Australian} climate data},
+  year      = {2001},
+  issn      = {1364-8152},
+  month     = jun,
+  number    = {4},
+  pages     = {309--330},
+  volume    = {16},
+  doi       = {10.1016/s1364-8152(01)00008-1},
+  publisher = {Elsevier BV},
+}
+
+
+@Article{Venaelaeinen2002,
+  author  = {Venäläinen, Ari and Heikinheimo, Martti},
+  journal = {Physics and Chemistry of the Earth, Parts A/B/C},
+  title   = {Meteorological data for agricultural applications},
+  year    = {2002},
+  issn    = {1474-7065},
+  number  = {23},
+  pages   = {1045--1050},
+  volume  = {27},
+  doi     = {10.1016/S1474-7065(02)00140-7},
+  url     = {http://www.sciencedirect.com/science/article/pii/S1474706502001407},
+}
+
+@Article{Ramachandra2007,
+  author   = {Ramachandra, T. V. and Shruthi, B. V.},
+  journal  = {Renewable and Sustainable Energy Reviews},
+  title    = {Spatial mapping of renewable energy potential},
+  year     = {2007},
+  issn     = {1364-0321},
+  number   = {7},
+  pages    = {1460--1480},
+  volume   = {11},
+  doi      = {10.1016/j.rser.2005.12.002},
+  keywords = {Renewable energy, Spatial analysis, Solar energy, Wind energy, Hydroenergy, Bioenergy, Energy demand, Energy potential, Bioenergy status, GIS},
+  url      = {http://www.sciencedirect.com/science/article/pii/S1364032106000050},
+}
+
+@Article{Alcoforado2009,
+  author   = {Alcoforado, Maria-João and Andrade, Henrique and Lopes, António and Vasconcelos, João},
+  journal  = {Landscape and Urban Planning},
+  title    = {Application of climatic guidelines to urban planning},
+  year     = {2009},
+  issn     = {0169-2046},
+  number   = {1},
+  pages    = {56--65},
+  volume   = {90},
[…]
