+++ R/get_dpird_summaries.R
@@ -98,6 +98,7 @@
 #'   * observationsCount,
 #'   * observationsPercentage,
 #'   * panEvaporation,
+#'   * panEvaporation12AM,
 #'   * rainfall,
 #'   * relativeHumidity,
 #'   * relativeHumidityAvg,
@@ -433,6 +434,7 @@ get_dpird_summaries <- function(station_code,
       "erosion_condition_minutes",
       "erosion_condition_start_time",
       "errors",
+      "evapotranspiration",
       "evapotranspiration_short_crop",
       "evapotranspiration_tall_crop",
       "frost_condition_minutes",
@@ -442,6 +444,7 @@ get_dpird_summaries <- function(station_code,
       "observations_count",
       "observations_percentage",
       "pan_evaporation",
+      "pan_evaporation_12am",
       "rainfall",
       "relative_humidity_avg",
       "relative_humidity_max",
@@ -497,6 +500,7 @@ get_dpird_summaries <- function(station_code,
       "erosion_condition_minutes",
       "erosion_condition_start_time",
       "errors",
+      "et",
       "et_short_crop",
       "et_tall_crop",
       "frost_condition_minutes",
@@ -506,6 +510,7 @@ get_dpird_summaries <- function(station_code,
       "observations_count",
       "observations_percentage",
       "pan_evaporation",
+      "pan_evaporation_12am",
       "rainfall",
       "rh_avg",
       "rh_tmax",

[…]
