+++ DESCRIPTION
@@ -19,8 +19,8 @@ Authors@R: c(
            comment = c(ORCID = "0000-0003-2172-5265")),
     person("Hugh", "Parsonage", , "hugh.parsonage@gmail.com", role = "aut",
            comment = c(ORCID = "0000-0003-4055-0835")),
-    person("Keith", "Pembleton", , "keith.pembleton@usq.edu.au", role = "aut",
-           comment = c(ORCID = "0000-0002-1896-4516")),
+    person("Keith", "Pembleton", , "keith.pembleton@usq.edu.au", role = "ctb",
+           comment = c(ORCID = "0000-0002-1896-4516", "Contributed code and ideas for original 'bomrang' package that was used in the creation of 'weatherOz'.")),
     person("Maëlle", "Salmon", , "maelle.salmon@yahoo.se", role = "ctb",
            comment = c(ORCID = "0000-0002-2815-0399", "Contributed to debugging a nasty little bug with CI where timezones caused tests to fail due to 'vcr' not recognising the\n                URL when run outside of Australia/Perth TZ! Suggested the use of `withr::local_timzeone()`, see <https://github.com/ropensci/weatherOz/commit/b052bf91973b8d7e147a39e8938405a64622634b>.")),
     person("Max", "Moldovan", , "max.moldovan@adelaide.edu.au", role = "ctb",
