### 2024-06-07 f274769
Ensure examples work and fix Adam's e-mail address
