+++ R/get_radar_imagery.R
@@ -26,6 +26,8 @@
 #' # Check availability radar imagery for Wollongong (radar_id = 3)
 #' imagery <- get_available_radar(radar_id = 3)
 #'
+#' imagery
+#'
 #' @family BOM
 #' @family metadata
 #'
