### 2024-06-07 6a06545
clean up DESCRIPTION formatting
