+++ DESCRIPTION
@@ -43,14 +43,14 @@ Description: Provides automated downloading, parsing and formatting of
     As well as the Bureau of Meteorology ('BOM') of the Australian
     government precis and coastal forecasts, agriculture bulletin data,
     and downloading and importing radar and satellite imagery files.
-    'DPIRD' weather data are accessed through public 'APIs' provided by 'DPIRD',
-    <https://www.agric.wa.gov.au/weather-api-20>, providing access to
-    weather station data from the 'DPIRD' weather station network.
-    Australia-wide weather data are based on data from the Australian
-    Bureau of Meteorology ('BOM') data and accessed through 'SILO' (Scientific
-    Information for Land Owners) <doi: 10.1016/S1364-8152(01)00008-1>.  'DPIRD'
-    data are made available under a Creative Commons Attribution 3.0 Licence
-    (CC BY 3.0 AU) license
+    'DPIRD' weather data are accessed through public 'APIs' provided by
+    'DPIRD', <https://www.agric.wa.gov.au/weather-api-20>, providing
+    access to weather station data from the 'DPIRD' weather station
+    network.  Australia-wide weather data are based on data from the
+    Australian Bureau of Meteorology ('BOM') data and accessed through
+    'SILO' (Scientific Information for Land Owners) <doi:
+    10.1016/S1364-8152(01)00008-1>.  'DPIRD' data are made available under
+    a Creative Commons Attribution 3.0 Licence (CC BY 3.0 AU) license
     <https://creativecommons.org/licenses/by/3.0/au/deed.en>.  SILO data
     are released under a Creative Commons Attribution 4.0 International
     licence (CC BY 4.0) <https://creativecommons.org/licenses/by/4.0/>.
