+++ R/get_precis_forecast.R
@@ -133,7 +133,7 @@ get_precis_forecast <- function(state = "AUS") {
 
 .parse_precis_forecast <- function(xml_url) {
 
-  op <- options(timeout = 480L)
+  op <- options(timeout = 600L)
   on.exit(options(op))
 
   # load the XML from ftp
