### 2024-06-07 #6717 @arfon
@JanLauGe, @rogerssam – many thanks for your reviews here! JOSS (and rOpenSci) relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
